% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/covariate_computation_functs.R
\name{accumulator_functs_enum}
\alias{accumulator_functs_enum}
\title{Accumulator function Enum
Function to return a simple enum of the accumulator functions}
\usage{
accumulator_functs_enum()
}
\value{

}
\description{
Accumulator function Enum
Function to return a simple enum of the accumulator functions
}
