% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{append_file_extension}
\alias{append_file_extension}
\title{Append file extension}
\usage{
append_file_extension(file_name, extension)
}
\arguments{
\item{file_name}{name of a file}

\item{extension}{extension}
}
\value{
the file name with the given extension
}
\description{
Appends a given file extension onto a file name (if not already present)
}
