% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s4_classes_methods.R
\name{apply,on_disc_matrix-method}
\alias{apply,on_disc_matrix-method}
\title{apply
apply a function to the rows or columns of an on_disc_matrix.}
\usage{
\S4method{apply}{on_disc_matrix}(X, MARGIN, FUN, chunk_size = 4000)
}
\arguments{
\item{X}{an on_disc_matrix}

\item{MARGIN}{apply to rows (1) or columns (2)}

\item{FUN}{a function to apply}

\item{chunk_size}{number of rows or columns to load at a time; default 4000}
}
\description{
apply
apply a function to the rows or columns of an on_disc_matrix.
}
