% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/covariate_computation_functs.R
\name{arguments_enum}
\alias{arguments_enum}
\title{Arguments Enum
Function to return a simple enum of the possible arguments to the accumulator functions}
\usage{
arguments_enum()
}
\value{

}
\description{
Arguments Enum
Function to return a simple enum of the possible arguments to the accumulator functions
}
