% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cell_cycle_genes.R
\name{assign_cell_cycle_phase}
\alias{assign_cell_cycle_phase}
\title{Assign cell cycle phases based on scores}
\usage{
assign_cell_cycle_phase(s_scores, g2m_scores)
}
\arguments{
\item{s_scores}{Numeric vector of S phase scores}

\item{g2m_scores}{Numeric vector of G2M phase scores}
}
\value{
Factor vector of phase assignments with levels c("G1", "S", "G2M", "Undecided")
}
\description{
Assign cell cycle phases based on scores
}
\keyword{internal}
