% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cell_cycle_genes.R
\docType{data}
\name{cc_genes_updated_2019}
\alias{cc_genes_updated_2019}
\title{Cell cycle genes: 2019 update}
\format{
A list with two components:
\describe{
\item{s_genes}{43 genes associated with S-phase}
\item{g2m_genes}{54 genes associated with G2M-phase}
}
}
\source{
\url{https://www.science.org/doi/abs/10.1126/science.aad0501}
}
\usage{
cc_genes_updated_2019
}
\description{
A list of genes used in cell-cycle regression, updated with 2019 symbols.
This matches Seurat's cc.genes.updated.2019 dataset exactly.
}
\section{Updated symbols}{

The following symbols were updated from the original gene sets:
\describe{
\item{s.genes}{
\itemize{
\item \emph{MCM2}: \emph{MCM7}
\item \emph{MLF1IP}: \emph{CENPU}
\item \emph{RPA2}: \emph{POLR1B}
\item \emph{BRIP1}: \emph{MRPL36}
}
}
\item{g2m.genes}{
\itemize{
\item \emph{FAM64A}: \emph{PIMREG}
\item \emph{HN1}: \emph{JPT1}
}
}
}
}

\examples{
# Load the cell cycle gene sets
data("cc_genes_updated_2019", package = "ondisc")
length(cc_genes_updated_2019$s_genes)   # 43 S-phase genes
length(cc_genes_updated_2019$g2m_genes) # 54 G2M-phase genes

}
\concept{data}
\keyword{datasets}
