% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{combine_multimodal_dataframes}
\alias{combine_multimodal_dataframes}
\title{Combine multimodal dataframes}
\usage{
combine_multimodal_dataframes(df_list, modality_names)
}
\arguments{
\item{df_list}{list of dfs}

\item{modality_names}{names of the modalities}
}
\value{
a combined dataframe containing all the cell-specific covariates
}
\description{
Combine multimodal dataframes
}
