% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{combine_results}
\alias{combine_results}
\title{Combine results}
\usage{
combine_results(res_list)
}
\arguments{
\item{res_list}{a list of results}
}
\value{
a vector or matrix containing the combined results
}
\description{
Takes a list of results (of the same type, i.e. vector or matrix) and combines them.
}
