% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_functions.R
\name{compare_Mat_on_disc_extract}
\alias{compare_Mat_on_disc_extract}
\title{Compare R sparse Matrix object to on_disc object on extract}
\usage{
compare_Mat_on_disc_extract(Mat, on_disc_mat, col_idxs, row_idxs)
}
\arguments{
\item{Mat}{an R sparse matrix object}

\item{on_disc_mat}{an on_disc_matrix}

\item{col_idxs}{column indices}

\item{row_idxs}{row indices}
}
\description{
Takes a sparse R matrix, on_disc_matrix, vector of column indexes, and vector of row indexes; verifies that the row, column, and row-column subsets match.
}
