% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ooc_pca.R
\name{compute_ooc_pca}
\alias{compute_ooc_pca}
\title{Compute out-of-core PCA on an \code{ondisc} matrix}
\usage{
compute_ooc_pca(covariate_odm)
}
\arguments{
\item{covariate_odm}{a \code{covariate_ondisc_matrix} object}
}
\description{
Compute out-of-core PCA on an \code{ondisc} matrix
}
