% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/low_level_initialize_transpose.R
\name{convert_mtx_to_csc}
\alias{convert_mtx_to_csc}
\title{Convert mtx file to compressed sparse column}
\usage{
convert_mtx_to_csc(mtx_fp, h5_loc, n_rows_with_comments, chunk_size = 1e+05)
}
\arguments{
\item{mtx_fp}{file path to mtx file}

\item{h5_loc}{file path to (already-initialized) h5 file}

\item{n_rows_with_comments}{number of rows with comments in the mtx file; the first n_rows_with_comments + 1 rows will be skipped.}

\item{chunk_size}{chunk size for the read_delim_chunked function; default 1e5}
}
\value{
NULL.
}
\description{
Indexing note: the input (.mtx) uses 1-based indexing for cell and gene. This function saves the (unchanged) expression data; the gene indexes, which are converted to 0-based indexing; and the row and column pointers, which are saved using 1-based indexing.
}
