% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/high_level_matrix_ops.R
\name{convert_to_seurat_object}
\alias{convert_to_seurat_object}
\title{Convert to Seurat object}
\usage{
convert_to_seurat_object(
  x,
  project = "CreateSeuratObject",
  assay = "RNA",
  cell_covariate_matrix = NULL
)
}
\arguments{
\item{x}{an on_disc_matrix}

\item{project}{(default "CreateSeuratObject") project name of the Seurat object}

\item{assay}{(default "RNA") name of the initial assay}

\item{cell_covariate_matrix}{(default none) cell-level covariate matrix}
}
\value{
a Seurat object
}
\description{
Converts an on_disc_matrix to a Suerat object.
}
