% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s4_classes_methods.R
\docType{class}
\name{covariate_ondisc_matrix-class}
\alias{covariate_ondisc_matrix-class}
\alias{covariate_ondisc_matrix}
\title{covariate_ondisc_matrix class constructor}
\value{
a covariate_ondisc_matrix object
}
\description{
A covariate_ondisc_matrix (short for "covariate ondisc_matrix") is an object that stores and ondisc_matrix, along with its cell-specific and feature-specific covariate matrices.
}
\section{Slots}{

\describe{
\item{\code{ondisc_matrix}}{an ondisc_matrix.}

\item{\code{cell_covariates}}{a data frame on cell covariates.}

\item{\code{feature_covariates}}{a data frame of feature covariates.}
}}

