% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s4_classes.R
\name{covariate_ondisc_matrix}
\alias{covariate_ondisc_matrix}
\title{\code{covariate_ondisc_matrix} class}
\usage{
covariate_ondisc_matrix(ondisc_matrix, cell_covariates, feature_covariates)

covariate_ondisc_matrix(ondisc_matrix, cell_covariates, feature_covariates)
}
\arguments{
\item{ondisc_matrix}{an \code{ondisc_matrix}.}

\item{cell_covariates}{a data frame storing the cell-specific covariates.}

\item{feature_covariates}{a data frame storing the feature-specific covariates.}
}
\value{
a \code{covariate_ondisc_matrix}.
}
\description{
A \code{covariate_ondisc_matrix} stores an \code{ondisc_matrix}, along with cell-specific and feature-specific covariate matrices.

Construct a \code{covariate_ondisc_matrix} by passing an \code{ondisc_matrix}, along with its associated \code{cell_covariates} and \code{feature_covariates}.
}
\section{Slots}{

\describe{
\item{\code{ondisc_matrix}}{an ondisc_matrix.}

\item{\code{cell_covariates}}{a data frame of cell covariates.}

\item{\code{feature_covariates}}{a data frame of feature covariates.}
}}

\examples{
# Install the `ondiscdata` package before running the examples.
# install.packages("devtools")
# devtools::install_github("Katsevich-Lab/ondiscdata")

# Load odm from package
odm_fp <- system.file("extdata", "odm/gene/matrix.odm", package = "ondiscdata")
metadata_fp <- system.file("extdata", "odm/gene/metadata.rds", package = "ondiscdata")
odm <- read_odm(odm_fp, metadata_fp)
odm_subset <- odm[1:10,]

covariate_odm <- covariate_ondisc_matrix(ondisc_matrix = get_ondisc_matrix(odm_subset),
                                         cell_covariates = get_cell_covariates(odm_subset),
                                         feature_covariates = get_feature_covariates(odm_subset))
}
