% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_functions.R
\name{create_new_directory}
\alias{create_new_directory}
\title{Create new directory}
\usage{
create_new_directory()
}
\value{
file path to a unique, empty directory
}
\description{
Creates a unique directory; returns a file path to that directory.
}
