% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_from_cellranger.R
\name{create_odm_from_cellranger}
\alias{create_odm_from_cellranger}
\title{Create ODM from cellranger}
\usage{
create_odm_from_cellranger(
  directories_to_load,
  directory_to_write,
  write_cellwise_covariates = TRUE,
  chunk_size = 1000L,
  compression_level = 3L,
  grna_target_data_frame = NULL
)
}
\arguments{
\item{directories_to_load}{directories containing the expression data}

\item{directory_to_write}{directory in which to write the initialized \code{.odm} files and \code{cellwise_covariates.rds} file}

\item{write_cellwise_covariates}{boolean indicating whether to write the cellwise covariates to disk in \code{.rds} format}

\item{chunk_size}{chunk size to use in the backing HDF5 file}

\item{compression_level}{compression level to use in the backing HDF5 file}

\item{grna_target_data_frame}{(optional) a data frame mapping each gRNA ID to its target. Relevant only if a CRISPR modality is present within the data. (See note.)}
}
\value{
list containing the odm object(s) and cellwise covariates
}
\description{
Create ODM from cellranger
}
\examples{
library(sceptredata)
directories_to_load <- paste0(
 system.file("extdata", package = "sceptredata"),
 "/highmoi_example/gem_group_", c(1, 2)
)
directory_to_write <- tempdir()
out_list <- create_odm_from_cellranger(
  directories_to_load = directories_to_load,
  directory_to_write = directory_to_write,
)
}
