% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_data_from_r_matrices.R
\name{create_odm_from_r_matrix}
\alias{create_odm_from_r_matrix}
\title{Create ODM from R matrix}
\usage{
create_odm_from_r_matrix(
  mat,
  file_to_write,
  chunk_size = 1000L,
  compression_level = 3L,
  integer_id = 0L
)
}
\arguments{
\item{mat}{an R matrix of class "dgRMatrix"}

\item{file_to_write}{name of the file in which to write the backing \code{.odm} file}

\item{chunk_size}{chunk size to use in the backing HDF5 file}

\item{compression_level}{compression level to use in the backing HDF5 file}

\item{integer_id}{ID used to identify the \code{.odm} file}
}
\value{
an odm object representing the data
}
\description{
Create ODM from R matrix
}
\examples{
set.seed(4)
n_row <- 500
n_col <- 800
v <- rpois(n = n_row * n_col, lambda = 1)
v[rbinom(n = n_row * n_col, size = 1, prob = 0.8) == 1] <- 0L
m <- matrix(data = v, nrow = n_row, ncol = n_col)
rownames(m) <- paste0("gene_", seq(1L, n_row))
mat <- as(m, "RsparseMatrix")
odm <- create_odm_from_r_matrix(mat, paste0(tempdir(), "gene.odm"))
}
