% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/high_level_initialize.R
\name{create_on_disc_matrix_from_10x_mtx}
\alias{create_on_disc_matrix_from_10x_mtx}
\title{Create on_disc_matrix from 10x .mtx file}
\usage{
create_on_disc_matrix_from_10x_mtx(
  mtx_fp,
  barcode_fp,
  features_fp,
  on_disc_dir,
  file_name = NULL
)
}
\arguments{
\item{mtx_fp}{file path to the .mtx file}

\item{barcode_fp}{file path to the barcode.tsv file.}

\item{features_fp}{file path to the features.tsv file.}

\item{on_disc_dir}{directory in which to store the initialized HDF5 object}

\item{file_name}{(optional) name of the .h5 file in which data are saved; defaults to on_disc_matrix_x.h5, where x is an integer.}
}
\value{
an on_disc_matrix object
}
\description{
This function creates an on_disc_matrix from a 10x .mtx file. The name of the created file (by default) is on_disc_matrix_x.h5, where x is an integer. The integer starts at 1; if there already exists a file called on_disc_matrix_1.h5, then the data will be saved in on_disc_matrix_2.h5, and so on. To override this behavior, explicitly specify the file name (WITHOUT the .h5 extension).
}
\examples{
mtx_fp <- system.file("extdata", "matrix_1.mtx", package = "ondisc")
barcode_fp <- system.file("extdata", "barcodes_1.tsv", package = "ondisc")
features_fp <- system.file("extdata", "features_1.tsv", package = "ondisc")
on_disc_dir <- system.file("extdata", package = "ondisc")
odm_fp <- system.file("extdata", "example.h5", package = "ondisc")
if (odm_fp != "") file.remove(odm_fp)
exp_mat <- create_on_disc_matrix_from_10x_mtx(mtx_fp,
barcode_fp,
features_fp,
on_disc_dir,
"example")
}
