% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/high_level_initialize.R
\name{create_ondisc_matrix_from_h5_list}
\alias{create_ondisc_matrix_from_h5_list}
\title{Create \code{ondisc_matrix} from h5}
\usage{
create_ondisc_matrix_from_h5_list(h5_list, odm_fp, progress = TRUE)
}
\arguments{
\item{h5_list}{a list of .h5 files to convert into \code{ondisc_matrix} format; the .h5 file should have the same features measured on different sets of cells.}

\item{odm_fp}{location to write the ondisc matrix to disk}

\item{progress}{progress (optional; default FALSE) print progress messages?}
}
\value{
A covariate_ondisc_matrix.
}
\description{
Creates an \code{ondisc_matrix} from a list of .h5 files storing single-cell expression data.
}
\examples{
\dontrun{
# ensure example .h5 data from "crisprdata" package are available
devtools::install_github("timothy-barry/crisprdata")
# get file paths to three .h5 files; these files contain different cells measured on the same genes
f_names <- c("GSM3722728_K562-dCas9-KRAB_5K-sgRNAs_Batch-4_2_filtered_gene_bc_matrices_h5.h5",
"GSM3722727_K562-dCas9-KRAB_5K-sgRNAs_Batch-4_1_filtered_gene_bc_matrices_h5.h5",
"GSM3722729_K562-dCas9-KRAB_5K-sgRNAs_Batch-1_1_filtered_gene_bc_matrices_h5.h5")
h5_list <- system.file("extdata", f_names, package = "crisprdata")
tempfile <- create_new_directory()
odm_fp <- paste0(tempfile, "/expression_odm")
# initialize a directory to store ondisc matrix
storage_dir <- create_new_directory()
# create the ondisc matrix
odm_plus_covariates_list <- create_ondisc_matrix_from_h5(h5_list, storage_dir)
}
}
