% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/high_level_initialize.R
\name{create_ondisc_matrix_from_mtx}
\alias{create_ondisc_matrix_from_mtx}
\title{Create an ondisc_matrix from a .mtx file.}
\usage{
create_ondisc_matrix_from_mtx(
  mtx_fp,
  barcodes_fp,
  features_fp,
  n_gb_per_chunk = 4,
  on_disc_dir = NULL,
  file_name = NULL
)
}
\arguments{
\item{mtx_fp}{file path to a .mtx file storing the expression data. The .mtx file can represent either an integer matrix or a logical (i.e., binary) matrix. If the .mtx file contains only two columns (after the initial three-column row of metadata), then the .mtx file is assumed to represent a logical matrix.}

\item{barcodes_fp}{file path to the .tsv file containing the cell barcodes.}

\item{features_fp}{file path to the features.tsv file. The first column (required) should contain the feature IDs (e.g., ENSG00000186092), and the second column (optional) should contain the human-readable feature names (e.g., OR4F5). Any subsequent columns are discarded.}

\item{n_gb_per_chunk}{(optional) amount of data (in GB) to process per chunk. Defaults to 4 GB.}

\item{on_disc_dir}{(optional) directory in which to store the on-disk portion of the ondisc_matrix. Defaults to the directory in which the .mtx file is located.}

\item{file_name}{(optional) name of the file in which to store the .h5 data on-disk. Defaults to ondisc_matrix_x.h5, where x is a unique integer starting at 1.}
}
\value{

}
\description{
Initializes an ondisc_matrix object from a .mtx file, a features.tsv file, and a barcodes.tsv file. Returns an ondisc_matrix R object along with cell-specific and feature-specific covariate matrices.
}
