% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{decrement_idxs}
\alias{decrement_idxs}
\title{decrement a vector of indexes}
\usage{
decrement_idxs(idxs)
}
\arguments{
\item{idxs}{an integer vector of indexes}
}
\description{
decrement a vector of indexes
}
