% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s4_class.R
\name{dim,odm-method}
\alias{dim,odm-method}
\title{dim}
\usage{
\S4method{dim}{odm}(x)
}
\arguments{
\item{x}{an object of class \code{odm}}
}
\value{
the dimension of the \code{odm}
}
\description{
Returns the dimension (i.e., number of features by number of cells) of an \code{odm}.
}
