% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s4_classes_methods.R
\name{dim,on_disc_matrix-method}
\alias{dim,on_disc_matrix-method}
\title{dim
Print dimension of on_disc_matrix}
\usage{
\S4method{dim}{on_disc_matrix}(x)
}
\arguments{
\item{x}{an on_disc_matrix}
}
\value{
an integer vector containing the dimension of the matrix
}
\description{
dim
Print dimension of on_disc_matrix
}
