% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s4_class.R
\name{dimnames,odm-method}
\alias{dimnames,odm-method}
\title{dimnames}
\usage{
\S4method{dimnames}{odm}(x)
}
\arguments{
\item{x}{an object of class \code{odm}}
}
\value{
the dimnames of the \code{odm}
}
\description{
Returns the dimnames of an \code{odm}.
}
