% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/covariate_computation_functs.R
\name{evaluate_grammar}
\alias{evaluate_grammar}
\title{Evaluate grammar}
\usage{
evaluate_grammar(symbol_name, grammar)
}
\arguments{
\item{symbol_name}{name of a symbol}

\item{grammar}{an environment representing a grammar}
}
\value{
the value of the symbol
}
\description{
Evaluates the grammar given a symbol name
}
