% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s4_subset_extract.R
\name{extract-odm}
\alias{extract-odm}
\alias{[[,ondisc_matrix,missing,missing-method}
\alias{[[,ondisc_matrix,ANY,missing-method}
\alias{[[,ondisc_matrix,missing,ANY-method}
\alias{[[,ondisc_matrix,ANY,ANY-method}
\alias{[[,covariate_ondisc_matrix,ANY,ANY-method}
\alias{[[,covariate_ondisc_matrix,missing,ANY-method}
\alias{[[,covariate_ondisc_matrix,ANY,missing-method}
\alias{[[,covariate_ondisc_matrix,missing,missing-method}
\alias{[[,multimodal_ondisc_matrix,ANY,ANY-method}
\title{Pull a submatrix into memory using the \code{[[} operator.}
\usage{
\S4method{[[}{ondisc_matrix,missing,missing}(x, i, j)

\S4method{[[}{ondisc_matrix,ANY,missing}(x, i, j)

\S4method{[[}{ondisc_matrix,missing,ANY}(x, i, j)

\S4method{[[}{ondisc_matrix,ANY,ANY}(x, i, j)

\S4method{[[}{covariate_ondisc_matrix,ANY,ANY}(x, i, j)

\S4method{[[}{covariate_ondisc_matrix,missing,ANY}(x, i, j)

\S4method{[[}{covariate_ondisc_matrix,ANY,missing}(x, i, j)

\S4method{[[}{covariate_ondisc_matrix,missing,missing}(x, i, j)

\S4method{[[}{multimodal_ondisc_matrix,ANY,ANY}(x, i, j)
}
\arguments{
\item{x}{an \code{ondisc_matrix} object.}

\item{i}{a vector (numeric, logical, or character) indicating features to pull into memory.}

\item{j}{a vector (numeric, logical, or character) indicating cells to pull into memory.}
}
\value{
a matrix (as implemented by the Matrix package).
}
\description{
Apply the \code{[[} operator to an \code{ondisc_matrix} or \code{covariate_ondisc_matrix} to pull a submatrix into memory. You can pass logical, character,
or numeric vectors to \code{[[}; character vectors are assumed to refer to feature IDs (for rows) and cell barcodes
(for columns).
}
\details{
One can remember the difference between \code{[} and \code{[[} by recalling R lists: \code{[} is used to subset a list, and
\code{[[} is used to access elements stored \emph{inside} a list. Similarly, \code{[} is used to subset an \code{ondisc_matrix}, and
\code{[[} is used to access a submatrix usable within R.
}
