% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{find_contig_subseqs}
\alias{find_contig_subseqs}
\title{find contiguous subsequences}
\usage{
find_contig_subseqs(v)
}
\arguments{
\item{v}{an integer vector}
}
\description{
find contiguous subsequences
}
