% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/low_level_initialize_transpose.R
\name{generate_on_disc_matrix_name}
\alias{generate_on_disc_matrix_name}
\title{Generate on disc_matrix_name}
\usage{
generate_on_disc_matrix_name(on_disc_dir)
}
\arguments{
\item{on_disc_dir}{directory in which to store the on_disc_matrix.}
}
\value{
a new name for an on_disc_matrix.
}
\description{
Generates the name of an on_disc_matrix object given a directory. This function searches for files named on_disc_matrix_\emph{.h5 in the specified directory. If none exists, it returns on_disc_matrix_1.h5. Else, it returns n_disc_matrix_}.h5 with a unique integer in place of *.
}
