% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s4_other_verbs.R
\name{get-covariates}
\alias{get-covariates}
\alias{get_cell_covariates}
\alias{get_feature_covariates}
\alias{get_cell_covariates,ondisc_matrix-method}
\alias{get_feature_covariates,ondisc_matrix-method}
\alias{get_cell_covariates,metadata_ondisc_matrix-method}
\alias{get_feature_covariates,metadata_ondisc_matrix-method}
\alias{get_cell_covariates,multimodal_ondisc_matrix-method}
\alias{get_feature_covariates,multimodal_ondisc_matrix-method}
\title{Get covariate matrices.}
\usage{
get_cell_covariates(x)

get_feature_covariates(x)

\S4method{get_cell_covariates}{ondisc_matrix}(x)

\S4method{get_feature_covariates}{ondisc_matrix}(x)

\S4method{get_cell_covariates}{metadata_ondisc_matrix}(x)

\S4method{get_feature_covariates}{metadata_ondisc_matrix}(x)

\S4method{get_cell_covariates}{multimodal_ondisc_matrix}(x)

\S4method{get_feature_covariates}{multimodal_ondisc_matrix}(x)
}
\arguments{
\item{x}{an object of class \code{ondisc_matrix}, \code{metadata_ondisc_matrix}, or \code{multimodal_ondisc_matrix}.}
}
\value{
a covariate matrix (in data frame form)
}
\description{
Functions to get cell-specific or feature-specific covariate matrices from \code{ondisc} objects.
}
