% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/low_level_initialize.R
\name{get_accumulator_funct_arg_list}
\alias{get_accumulator_funct_arg_list}
\title{Get accumulator funct and arg list}
\usage{
get_accumulator_funct_arg_list(terminal_symbol)
}
\arguments{
\item{terminal_symbol}{a terminal symbol}
}
\value{
a list containing (i) the name of and (ii) the arguments to pass to the accumulator function.
}
\description{
Get accumulator funct and arg list
}
