% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/high_level_matrix_ops.R
\name{get_cell_barcodes}
\alias{get_cell_barcodes}
\alias{get_gene_ids}
\alias{get_gene_names}
\title{Get names}
\usage{
get_cell_barcodes(x)

get_gene_ids(x)

get_gene_names(x)
}
\arguments{
\item{x}{An on_disc_matrix.}
}
\value{
The cell barcodes of this on-disc matrix.

The cell gene ids of this on-disc matrix.

The gene names of this on-disc matrix.
}
\description{
Functions to extract cell barcodes, gene ids, and gene names from an on_disc_matrix.
}
