% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qc_helpers.R
\name{get_cells_with_low_p_mito}
\alias{get_cells_with_low_p_mito}
\title{Get cells with low proportion mitochondrial reads}
\usage{
get_cells_with_low_p_mito(covariate_odm, p_mito_thresh = 0.1)
}
\arguments{
\item{covariate_odm}{a \code{covariate_ondisc_matrix} object}

\item{p_mito_thresh}{proportion mitochondrial reads threshold}
}
\value{
a vector of cell barcodes
}
\description{
Obtains the cells with proportion mitochondrial reads below a given threshold.
}
\examples{
# Install the `ondiscdata` package to run the examples.
# devtools::install_github("timothy-barry/ondiscdata")

odm_fp <- system.file("extdata", "odm/gene/matrix.odm", package = "ondiscdata")
metadata_fp <- system.file("extdata", "odm/gene/metadata.rds", package = "ondiscdata")
cov_odm <- read_odm(odm_fp, metadata_fp)
cells_to_keep <- get_cells_with_low_p_mito(cov_odm)
}
