% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qc_helpers.R
\name{get_cells_with_moderate_seq_depth}
\alias{get_cells_with_moderate_seq_depth}
\title{Get cells with moderate sequencing depth}
\usage{
get_cells_with_moderate_seq_depth(
  covariate_odm,
  lower_percentile = 0.03,
  upper_percentile = 0.97
)
}
\arguments{
\item{covariate_odm}{a \code{covariate_ondisc_matrix} object}

\item{lower_percentile}{lower library size percentile}

\item{upper_percentile}{upper library size percentile}
}
\value{
a vector of cell barcodes
}
\description{
Obtains the cells with a library size greater than \code{lower_percentile} and less than \code{upper_percentile}.
}
\examples{
# Install the `ondiscdata` package before running the examples.
# install.packages("devtools")
# devtools::install_github("Katsevich-Lab/ondiscdata")

odm_fp <- system.file("extdata", "odm/gene/matrix.odm", package = "ondiscdata")
metadata_fp <- system.file("extdata", "odm/gene/metadata.rds", package = "ondiscdata")
cov_odm <- read_odm(odm_fp, metadata_fp)
cells_to_keep <- get_cells_with_moderate_seq_depth(cov_odm)
}
