% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{get_chunks_sequence}
\alias{get_chunks_sequence}
\title{Get a chunks sequence}
\usage{
get_chunks_sequence(n_items, chunk_size)
}
\arguments{
\item{n_items}{number of items (genes or cells)}

\item{chunk_size}{size of chunk}
}
\value{
a list of indices v; for entry i, (v(i) + 1):(v(i+1)) provides a range.
}
\description{
Generates a sequence of cells/genes to load given a chunk size and the number of cells/genes.
}
