% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/high_level_matrix_ops.R
\name{get_dim}
\alias{get_dim}
\title{Get dim}
\usage{
get_dim(x)
}
\arguments{
\item{x}{an ondisc_matrix object}
}
\value{
the dimension
}
\description{
Returns the dimension of an on_disc matrix.
}
