% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s4_class.R
\name{get_feature_ids}
\alias{get_feature_ids}
\title{Get feature IDs}
\usage{
get_feature_ids(odm)
}
\arguments{
\item{odm}{an object of class \code{odm}}
}
\value{
a character vector containing the feature IDs
}
\description{
Returns the feature IDs of an ODM.
}
