% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/high_level_initialize_helper.R
\name{get_features_metadata}
\alias{get_features_metadata}
\title{Get metadata for features.tsv file}
\usage{
get_features_metadata(features, bag_of_variables, is_file = TRUE)
}
\arguments{
\item{features}{file path to features.tsv file or a data frame giving the names of the features.}

\item{bag_of_variables}{the bag of variables to which to add the mt_genes logical vector (if applicable)}

\item{is_file}{a boolean variable indicating: TRUE, features is a file path or FALSE, features is a data frame. Defualt TRUE.}
}
\value{
a list containing elements feature_names (logical), n_cols (integer), and whether MT genes are present (logical)
}
\description{
Gets metadata from a features.tsv file. As a side-effect, if MT genes are present, puts into the bag_of_variables a logical vector indicating the positions of those genes.
}
