% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/high_level_initialize_helper.R
\name{get_features_metadata_from_table}
\alias{get_features_metadata_from_table}
\title{Get metadata for features_df, a data frame giving the names of the features.}
\usage{
get_features_metadata_from_table(features_df, bag_of_variables = NULL)
}
\arguments{
\item{features_df}{a data frame giving the names of the features.}

\item{bag_of_variables}{the bag of variables to which to add the mt_genes logical vector (if applicable)}
}
\value{
a list containing elements feature_names (logical), n_cols (integer), and whether MT genes are present (logical)
}
\description{
Gets metadata from a features data frame features_df.
}
