% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/high_level_initialize_helper.R
\name{get_h5_barcodes}
\alias{get_h5_barcodes}
\title{Get h5 barcodes}
\usage{
get_h5_barcodes(h5_list)
}
\arguments{
\item{h5_list}{a vector of paths to the h5 file}
}
\value{
a list of barcodes for each h5 file
}
\description{
Get h5 barcodes
}
