% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/high_level_initialize_helper.R
\name{get_h5_cells_metadata}
\alias{get_h5_cells_metadata}
\title{Get h5 cells metadata}
\usage{
get_h5_cells_metadata(h5_list)
}
\arguments{
\item{h5_list}{a vector of paths to the h5 file}
}
\value{
a list containing (i) n_genes, (ii) n_cells, (iii) the number of
data points (i.e., fraction of entries that are zero)
}
\description{
Get h5 cells metadata
}
