% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/high_level_initialize_helper.R
\name{get_h5_features}
\alias{get_h5_features}
\title{Get h5 features}
\usage{
get_h5_features(h5_list)
}
\arguments{
\item{h5_list}{a vector of paths to the h5 file}
}
\value{
features_df data frame giving the names of the features. The first column (required) contains the feature IDs (e.g., ENSG00000186092), and the second column (optional) contains the human-readable feature names (e.g., OR4F5). Subsequent columns are discarded. Gene names starting with "MT-" are assumed to be mitochondrial genes and will be used to compute the p_mito covariate.
}
\description{
Get h5 features
}
