% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qc_helpers.R
\name{get_highly_expressed_features}
\alias{get_highly_expressed_features}
\title{Get highly expressed features}
\usage{
get_highly_expressed_features(covariate_odm, frac_expressed = 0.05)
}
\arguments{
\item{covariate_odm}{a \code{covariate_ondisc_matrix} object}

\item{frac_expressed}{a fraction; features expressed in this fraction (or more) of cells are returned}
}
\value{
the highly expressed features
}
\description{
Obtains the set of features expressed in a least \code{frac_expressed} of cells.
}
\examples{
# Install the `ondiscdata` package before running the examples.
# install.packages("devtools")
# devtools::install_github("Katsevich-Lab/ondiscdata")

odm_fp <- system.file("extdata", "odm/gene/matrix.odm", package = "ondiscdata")
metadata_fp <- system.file("extdata", "odm/gene/metadata.rds", package = "ondiscdata")
cov_odm <- read_odm(odm_fp, metadata_fp)
highly_expressed_features <- get_highly_expressed_features(cov_odm)
}
