% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qc_helpers.R
\name{get_highly_variable_features}
\alias{get_highly_variable_features}
\title{Get highly variable features}
\usage{
get_highly_variable_features(covariate_odm, n_features = 250)
}
\arguments{
\item{covariate_odm}{a \code{covariate_ondisc_matrix} object}

\item{n_features}{the number of features to extract}
}
\value{
the top \code{n_features} most variable features
}
\description{
Obtains the most variable features from a \code{covariate_ondisc_matrix}, as measured by coefficient of variation.
}
\examples{
# Install the `ondiscdata` package to run the examples.
# devtools::install_github("timothy-barry/ondiscdata")

odm_fp <- system.file("extdata", "odm/gene/matrix.odm", package = "ondiscdata")
metadata_fp <- system.file("extdata", "odm/gene/metadata.rds", package = "ondiscdata")
cov_odm <- read_odm(odm_fp, metadata_fp)
# first, subset to examine highly expressed features only
cov_odm_highly_exp <- cov_odm[get_highly_expressed_features(cov_odm),]
# now, extract the highly variable features
highly_variable_features <- get_highly_variable_features(cov_odm_highly_exp)
}
