% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_functions.R
\name{get_metadata_odm_list}
\alias{get_metadata_odm_list}
\title{Get metadata odm list}
\usage{
get_metadata_odm_list(mat_list, idx_start, temp_test_dir)
}
\arguments{
\item{mat_list}{a list of matrices}

\item{idx_start}{the starting index}

\item{temp_test_dir}{directory used by the tests}
}
\value{
a list of initialized metadata_odms
}
\description{
Get metadata odm list
}
