% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s4_other_verbs.R
\name{get_modality}
\alias{get_modality}
\title{Get modality}
\usage{
get_modality(multimodal_mat, modality_name)
}
\arguments{
\item{multimodal_mat}{a multimodal_ondisc_matrix object.}

\item{modality_name}{name of modality to extract.}
}
\value{
the requested modality
}
\description{
Obtains a given modality from a multimodal_ondisc_matrix object by name.
}
