% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/high_level_initialize_helper.R
\name{get_mtx_metadata}
\alias{get_mtx_metadata}
\title{Get mtx metadata}
\usage{
get_mtx_metadata(mtx_fp, n_rows_with_comments)
}
\arguments{
\item{mtx_fp}{filepath to the mtx file}

\item{n_rows_with_comments}{number of rows with comments (at top of file)}
}
\value{
a list containing (i) n_genes, (ii) n_cells, (iii) the sparsity (i.e., fraction of entries that are zero), (iv) (TRUE/FALSE) matrix is logical
}
\description{
Get mtx metadata
}
