% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/high_level_initialize_helper.R
\name{get_n_rows_with_comments_mtx}
\alias{get_n_rows_with_comments_mtx}
\title{get n rows with comments mtx}
\usage{
get_n_rows_with_comments_mtx(mtx_fp)
}
\arguments{
\item{mtx_fp}{file path to mtx}
}
\value{
the number of rows in the mtx with comments
}
\description{
get n rows with comments mtx
}
