% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/high_level_matrix_ops.R
\name{get_names}
\alias{get_names}
\title{Get names helper function}
\usage{
get_names(x, name_to_get)
}
\arguments{
\item{x}{an ondisc_matrix}

\item{name_to_get}{the names to extract from x (one of cell_barcodes, feature_ids, feature_names)}
}
\value{
a character vector containing the requested names
}
\description{
Get names helper function
}
