% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_functions.R
\name{get_random_subset}
\alias{get_random_subset}
\title{get_random_subset}
\usage{
get_random_subset(n)
}
\arguments{
\item{n}{an integer}
}
\value{
a random subset of 1, 2, ..., n.
}
\description{
get_random_subset
}
