% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_functions.R
\name{get_simulation_data_fps}
\alias{get_simulation_data_fps}
\title{Get simulation data filepaths}
\usage{
get_simulation_data_fps(data_dir, idx)
}
\arguments{
\item{data_dir}{directory in which the simulation objects are stored}

\item{idx}{an index}
}
\value{
a character vector containing file paths to the simulation data
}
\description{
Get file paths to simulation objects given an index.
}
