% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/high_level_initialize_helper.R
\name{get_string_arrays}
\alias{get_string_arrays}
\title{Get string}
\usage{
get_string_arrays(barcodes_fp, features_fp, bag_of_variables, barcode_suffixes)
}
\arguments{
\item{barcodes_fp}{path to barcodes file(s)}

\item{features_fp}{path the features file}

\item{bag_of_variables}{stores the features_metadata list}

\item{barcode_suffixes}{a vector of suffix that appended to the barcodes to for each input barcode_fp. If NULL, append file index.}
}
\value{
list containing (i) cell barcodes, (ii) feature IDs, and (iii) feature names (NA_character_ if absent)
}
\description{
Get string
}
