% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/covariate_computation_functs.R
\name{get_terminal_acc_and_args}
\alias{get_terminal_acc_and_args}
\title{Get terminal accumulator and args}
\usage{
get_terminal_acc_and_args(terminal_symbol)
}
\arguments{
\item{terminal_symbol}{a terminal symbol}
}
\value{
a list containing (i) the name of the function, and (ii) the ordered argument names.
}
\description{
Returns the accumulator function and ordered argument set for a given terminal symbol.
}
