% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/covariate_computation_functs.R
\name{get_terminals_for_covariate}
\alias{get_terminals_for_covariate}
\title{Get nonterminals for covariate}
\usage{
get_terminals_for_covariate(covariate, grammar)
}
\arguments{
\item{covariate}{name of covariate}

\item{grammar}{a grammar, as initialized by initialize_grammar.}
}
\value{
character vector of nonterminals
}
\description{
Obtains the nonterminals needed to compute a given covariate.
}
