% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_functions.R
\name{get_test_parameters}
\alias{get_test_parameters}
\title{Get test parameters}
\usage{
get_test_parameters(test_type)
}
\arguments{
\item{test_type}{"small" or "big," indicating which set of tests to run.}
}
\value{
a list of parameters: synthetic_data_dir, n_datasets, n_reps_per_dataset
}
\description{
Obtain the parameters for a run of test. The "small" test is lightweight and can easily be run on any machine on which ondisc is installed. The "large" test, by contrast, is heavier duty and takes a bit more effort to set up.
}
