% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s4_classes_methods.R
\name{head,on_disc_matrix-method}
\alias{head,on_disc_matrix-method}
\title{print first rows and columns of an on_disc_matrix}
\usage{
\S4method{head}{on_disc_matrix}(x)
}
\arguments{
\item{x}{an on_disc_mnatrix}
}
\description{
print first rows and columns of an on_disc_matrix
}
