% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{inc_cell_count_if_feature_condition}
\alias{inc_cell_count_if_feature_condition}
\title{increment count (if condition on feature holds)}
\usage{
inc_cell_count_if_feature_condition(
  acc_vect,
  feature_idxs,
  cell_idxs,
  umi_counts,
  bool_vect
)
}
\arguments{
\item{acc_vect}{an integer accumulator vector}

\item{feature_idxs}{0-based integer vector of feature indexes}

\item{cell_idxs}{0-based integer vector of cell_idxs}

\item{umi_counts}{integer vector of UMI counts}

\item{bool_vect}{a logical vector indicating whether to increment acc_vect; should be same length as n_features}
}
\description{
increment count (if condition on feature holds)
}
