% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{inc_count}
\alias{inc_count}
\title{increment count}
\usage{
inc_count(acc_vect, idxs, umi_counts)
}
\arguments{
\item{acc_vect}{an integer accumulator vector}

\item{idxs}{0-based integer vector of indexes}

\item{umi_counts}{integer vector of UMI counts}
}
\description{
increment count
}
