% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{inc_mean_sq_count}
\alias{inc_mean_sq_count}
\title{increment mean squared count}
\usage{
inc_mean_sq_count(acc_vect, idxs, umi_counts, n)
}
\arguments{
\item{acc_vect}{a numeric accumulator vector}

\item{idxs}{0-based integer vector of indexes (e.g., feature indexes)}

\item{umi_counts}{integer vector of UMI counts}

\item{n}{a double indicating the total number of elements to divide by (e.g., number of cells)}
}
\description{
increment mean squared count
}
