% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{inc_n_entries}
\alias{inc_n_entries}
\title{increment n entries}
\usage{
inc_n_entries(acc_vect, idxs)
}
\arguments{
\item{acc_vect}{an integer accumulator vector}

\item{idxs}{0-based integer vector of indexes}
}
\description{
increment n entries
}
