% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/low_level_initialize_transpose.R
\name{increment_indices}
\alias{increment_indices}
\title{Increment indices}
\usage{
increment_indices(old_indices, new_indices)
}
\arguments{
\item{old_indices}{A vector}

\item{new_indices}{the indices (optionally repeated) to increment by 1 in the vector}
}
\value{
the old_indices vector with appropriately updated values.
}
\description{
A simple helper function to increment the indices of a vector. (Note: re-write in C++)
}
