% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{index_h5_file}
\alias{index_h5_file}
\title{index into an h5 file}
\usage{
index_h5_file(
  file_name_in,
  p_name_in,
  idx_name_in,
  umi_counts_name_in,
  subset_vector,
  logical_mat
)
}
\arguments{
\item{file_name_in}{name of h5 file}

\item{p_name_in}{name of pointer}

\item{idx_name_in}{name of (minor) index}

\item{umi_counts_name_in}{name of umi counts}

\item{subset_vector}{the integer vector of indexes to extract}

\item{logical_mat}{boolean indicating whether the matrix is logical}
}
\description{
index into an h5 file
}
