% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s4_classes_methods.R
\docType{class}
\name{index_vec-class}
\alias{index_vec-class}
\title{index_vec class
A union of the numeric, logical, and character classes. Used for indexing of on_disc_matrices. Idea borrowed from Matrix package.}
\description{
index_vec class
A union of the numeric, logical, and character classes. Used for indexing of on_disc_matrices. Idea borrowed from Matrix package.
}
