% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/low_level_initialize.R
\name{initialize_accumulator}
\alias{initialize_accumulator}
\title{Initialize accumulator}
\usage{
initialize_accumulator(terminal_symbol, bag_of_variables)
}
\arguments{
\item{terminal_symbol}{the terminal symbol}

\item{bag_of_variables}{the bag of variables}
}
\value{
An initialized vector of the correct type and length
}
\description{
Initialize accumulator
}
