% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/covariate_computation_functs.R
\name{initialize_grammar}
\alias{initialize_grammar}
\title{Initialize grammar}
\usage{
initialize_grammar()
}
\value{
an environment representing the context-free grammar.
}
\description{
Initializes the context-free grammar used to compute the covariate matrices.
}
