% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/low_level_initialize.R
\name{initialize_h5_file_on_disk}
\alias{initialize_h5_file_on_disk}
\title{Initialize h5 file on-disk}
\usage{
initialize_h5_file_on_disk(
  h5_fp,
  mtx_metadata,
  features_metadata,
  barcodes_fp,
  features_fp
)
}
\arguments{
\item{h5_fp}{file path to the .h5 file to be initialized}

\item{mtx_metadata}{metadata of the .mtx file}

\item{features_metadata}{metadata of the features.tsv file}

\item{barcodes_fp}{file path to the barcodes.tsv file}

\item{features_fp}{file path to the features.tsv file}
}
\description{
Initialize the on-disk portion on an ondisc_matrix.
}
