% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s4_class.R
\name{initialize_odm_from_backing_file}
\alias{initialize_odm_from_backing_file}
\title{Initialize an ondisc matrix}
\usage{
initialize_odm_from_backing_file(odm_file)
}
\arguments{
\item{odm_file}{file path to the backing \code{.odm} file}
}
\value{
an object of class \code{odm}
}
\description{
Initializes an object of class \code{odm} from a backing .odm file.
}
\note{
The file path \code{odm_file} should be the fully qualified file path to the backing .odm file. In particular, \code{odm_file} should not contain the home directory symbol (\code{~}) or current directory symbol (\code{.}).
}
