% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/high_level_initialize.R
\name{internal_initialize_ondisc_matrix}
\alias{internal_initialize_ondisc_matrix}
\title{internal initialize ondisc_matrix}
\usage{
internal_initialize_ondisc_matrix(h5_file, logical_mat, underlying_dimension)
}
\arguments{
\item{h5_file}{an h5 file}

\item{logical_mat}{logical value indicating whether the matrix is logical}

\item{underlying_dimension}{length-two integer vector giving the dimension of the underlying matrix}
}
\value{
a correctly-initialized ondisc_matrix
}
\description{
An internal function for initializing an ondisc_matrix
}
