% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_functions.R
\name{load_on_disc_and_mat}
\alias{load_on_disc_and_mat}
\title{Load on disc matrix and R sparse matrix}
\usage{
load_on_disc_and_mat(data_dir, idx)
}
\arguments{
\item{data_dir}{simulation data directory}

\item{idx}{index}
}
\value{
a list containing the on_disc_matrix and the original sparse R matrix.
}
\description{
Load on disc matrix and R sparse matrix
}
