% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/covariate_computation_functs.R
\name{map_inputs_to_covariates}
\alias{map_inputs_to_covariates}
\title{Map inputs to covariates}
\usage{
map_inputs_to_covariates(mtx_metadata, features_metadata)
}
\arguments{
\item{mtx_metadata}{the mtx metadata}

\item{features_metadata}{the features.tsv metadata}
}
\value{
a list of two entries: feature_covariates and cell_covariates; each entry lists the covariates to compute
}
\description{
A function that maps the input metadata to the covariates to compute.
}
