% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{map_memory_to_disk_logical_matrix}
\alias{map_memory_to_disk_logical_matrix}
\title{map memory to disk}
\usage{
map_memory_to_disk_logical_matrix(
  file_name_in,
  m_cell_idxs,
  cell_idxs_name,
  n_features,
  m_row_ptr,
  f_row_ptr
)
}
\arguments{
\item{file_name_in}{name of h5 file}

\item{m_cell_idxs}{the cell indexes integer vector}

\item{cell_idxs_name}{name of the cell_idxs dataset}

\item{n_features}{number of features}

\item{m_row_ptr}{the memory row pointer}

\item{f_row_ptr}{the disk row pointer}
}
\description{
map memory to disk
}
