% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s4_classes_methods.R
\name{multimodal_ondisc_matrix}
\alias{multimodal_ondisc_matrix}
\title{Constructor for multimodal_ondisc_matrix}
\usage{
multimodal_ondisc_matrix(covariate_ondisc_matrix_list)
}
\arguments{
\item{covariate_ondisc_matrix_list}{a named list containing covariate_ondisc_matrices; the names are taken to be the names of the modalities}
}
\value{
a multimodal_ondisc_matrix
}
\description{
Constructs a multimodal_ondisc_matrix from a list of covariate_ondisc_matrix objects
}
\examples{
# NOTE: You must create the RDS files "expressions.rds" and
# "perturbations.rds" to run this example. Navigate to the help file of
# "create_ondisc_matrix_from_mtx" (via ?create_ondisc_matrix_from_mtx),
# and execute both code blocks.
expression_fp <- paste0(tempdir(), "/expressions.rds")
perturbations_fp <- paste0(tempdir(), "/perturbations.rds")
if (file.exists(expression_fp) && file.exists(perturbations_fp)) {
    expressions <- readRDS(expression_fp)
    perturbations <- readRDS(expression_fp)
    crispr_experiment <- multimodal_ondisc_matrix(list(expressions = expressions,
    perturbations = perturbations))
}
}
