% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/high_level_initialize_helper.R
\name{n_gb_to_n_entries}
\alias{n_gb_to_n_entries}
\title{n GB to entries}
\usage{
n_gb_to_n_entries(n_gb, logical_mtx)
}
\arguments{
\item{n_gb}{number of gigabytes to process per chunk}

\item{logical_mtx}{number of}
}
\value{

}
\description{
n GB to entries
}
