% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/post_load_processing.R
\name{normalize_by_regression}
\alias{normalize_by_regression}
\title{Normalize by regression}
\usage{
normalize_by_regression(
  covariate_odm,
  covariates = c("p_mito", "lg_n_nonzero"),
  offset = "lg_n_umis"
)
}
\arguments{
\item{covariate_odm}{a \code{covariate_odm} object}

\item{covariates}{a character vector listing the covariates within the cell covariate matrix to regress on}

\item{offset}{a character vector giving the variable within the cell covariate matrix to use as an offset}
}
\value{
a normalized \code{covariate_odm} object
}
\description{
Normalize by regression
}
\examples{
library(magrittr)
odm_fp <- system.file("extdata", "odm/gene/matrix.odm", package = "ondiscdata")
metadata_fp <- system.file("extdata", "odm/gene/metadata.rds", package = "ondiscdata")
odm <- read_odm(odm_fp, metadata_fp)
my_feats <- sample(get_highly_expressed_features(odm, 0.3), 100)
covariate_odm <- odm[my_feats,]
# add log-transformed n_umis, n_nonero
covariate_odm <- covariate_odm \%>\% mutate_cell_covariates(lg_n_umis = log(n_umis),
lg_n_nonzero = log(n_nonzero))
# regress on the default covariates, i.e. p_mito and lg_n_nonzero
covariate_odm_norm <- normalize_by_regression(covariate_odm)
# regress only on the default offset, lg_n_umis
covariate_odm_norm <- normalize_by_regression(covariate_odm, covariates = NULL)
}
