% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/high_level_matrix_ops.R
\name{on_disc_apply_across_chunks}
\alias{on_disc_apply_across_chunks}
\title{On disc apply across chunks}
\usage{
on_disc_apply_across_chunks(x, col_apply, chunk_function, chunk_size)
}
\arguments{
\item{x}{an on_disc matrix}

\item{col_apply}{(boolean) apply across columns (TRUE) or rows (FALSE)}

\item{chunk_function}{function to apply to each chunk}

\item{chunk_size}{size of the chunks to load}
}
\value{
a vector or matrix containing the results of applying the function to the on_disc matrix.
}
\description{
Applies a function across chunks of an on_disc_matrix.
}
