% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s4_classes_methods.R
\docType{class}
\name{on_disc_matrix-class}
\alias{on_disc_matrix-class}
\alias{on_disc_matrix}
\title{The on_disc_matrix class}
\description{
An on_disc_matrix represents a gene-by-cell expression matrix stored on disk.
}
\section{Slots}{

\describe{
\item{\code{h5_file}}{path to an expression matrix (created by one of the initialization functions in this package) stored on disk}

\item{\code{cell_subset}}{an integer vector storing the cells currently in use}

\item{\code{gene_subset}}{an integer vector storing the genes currently in use}
}}

