% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ondisc.R
\docType{package}
\name{ondisc-package}
\alias{ondisc}
\alias{ondisc-package}
\title{ondisc: Algorithms and data structures for large single-cell expression matrices}
\description{
Single-cell datasets are growing in size, posing challenges as well as opportunities to genomics researchers. `ondisc` is an R package for computing on large-scale single-cell data. `ondisc` enables the analysis of data out-of-core on a laptop or distributed across multiple nodes on a computing cluster. In both settings `ondisc` requires only a few gigabytes of memory, even if the input data are tens of gigabytes in size. `ondisc` is oriented toward single-cell differential expression, single-cell co-expresssion, and single-cell CRISPR screen analyses.
}
\author{
\strong{Maintainer}: Timothy Barry \email{tbarry@hsph.harvard.edu} (\href{https://orcid.org/0000-0002-4356-627X}{ORCID})

Authors:
\itemize{
  \item Eugene Katsevich \email{ekatsevi@wharton.upenn.edu} [thesis advisor]
}

}
