% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ondisc.R
\docType{package}
\name{ondisc}
\alias{ondisc}
\title{ondisc: A package for out-of-memory computing on single-cell data}
\description{
Single-cell datasets are large and are growing in size as sequencing costs drop. The ondisc package is designed to facilitate large-scale computing on single-cell expression data by providing access to expression matrices out-of-memory. ondisc is functional (i.e., all objects are persistent) and efficient (i.e., all algorithms are theoretically optimal in time).
}
