% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s4_classes.R
\docType{class}
\name{ondisc_matrix-class}
\alias{ondisc_matrix-class}
\alias{ondisc_matrix}
\title{\code{ondisc_matrix} class}
\description{
An \code{ondisc_matrix} represents a feature-by-cell expression matrix stored on-disk.
}
\details{
It is best to avoid interacting with the slots of an \code{ondisc_matrix} directly. Instead, use the functions
and operators provided by the package.
}
\section{Slots}{

\describe{
\item{\code{h5_file}}{path to an initialized .h5 file stored on-disk.}

\item{\code{logical_mat}}{logical value indicating whether the matrix is logical.}

\item{\code{cell_subset}}{integer vector recording the cells currently in use.}

\item{\code{feature_subset}}{integer vector recording the features currently in use.}

\item{\code{underlying_dimension}}{the dimension of the (unsubset) expression matrix.}
}}

