% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s4_classes.R
\name{ondisc_matrix}
\alias{ondisc_matrix}
\title{\code{ondisc_matrix} class}
\usage{
ondisc_matrix(
  h5_file = NA_character_,
  logical_mat = FALSE,
  underlying_dimension = NA_integer_,
  cell_subset = NA_integer_,
  feature_subset = NA_integer_,
  feature_ids = NA_character_,
  feature_names = NA_character_,
  cell_barcodes = NA_character_,
  odm_id = NA_integer_
)

ondisc_matrix(
  h5_file = NA_character_,
  logical_mat = FALSE,
  underlying_dimension = NA_integer_,
  cell_subset = NA_integer_,
  feature_subset = NA_integer_,
  feature_ids = NA_character_,
  feature_names = NA_character_,
  cell_barcodes = NA_character_,
  odm_id = NA_integer_
)
}
\arguments{
\item{h5_file}{location of backing .odm file on disk.}

\item{logical_mat}{boolean indicating whether matrix is logical}

\item{underlying_dimension}{dimension of underlying expression matrix}

\item{cell_subset}{integer vector indicating cells in use}

\item{feature_subset}{integet vector indicating features in use}

\item{feature_ids}{character vector of feature IDs}

\item{feature_names}{character vector of feature names}

\item{cell_barcodes}{character vector of cell barcodes}

\item{odm_id}{unique (with high probability) integer}
}
\value{
initialized \code{ondisc_matrix} object
}
\description{
An \code{ondisc_matrix} represents a feature-by-cell expression matrix stored on-disk.

Constructor function for \code{ondisc_matrix} class.
}
\details{
It is best to avoid interacting with the slots of an \code{ondisc_matrix} directly. Instead, use the functions
and operators provided by the package.
}
\examples{
# Use `ondiscdata` package for the examlpes, please install the package before running the examples
# install.packages("devtools")
# devtools::install_github("Katsevich-Lab/ondiscdata")

# Load odm from package
odm_fp <- system.file("extdata", "odm/gene/matrix.odm", package = "ondiscdata")
metadata_fp <- system.file("extdata", "odm/gene/metadata.rds", package = "ondiscdata")
odm <- read_odm(odm_fp, metadata_fp)
odm_subset <- odm[1:10,]

tempfile <- create_new_directory()
h5_file <- paste0(tempfile, "/expression_odm")
odm_output <- ondisc_matrix(h5_file = h5_file,
                     logical_mat = odm_subset@ondisc_matrix@logical_mat,
                     underlying_dimension = odm_subset@ondisc_matrix@underlying_dimension,
                     feature_ids = odm_subset@ondisc_matrix@feature_ids,
                     feature_names = odm_subset@ondisc_matrix@feature_names,
                     cell_barcodes = odm_subset@ondisc_matrix@cell_barcodes,
                     odm_id = sample(seq(0L, .Machine$integer.max), size = 1))
}
