% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{read_dimension}
\alias{read_dimension}
\title{Read the dimension of the ODM into memory}
\usage{
read_dimension(file_name_in)
}
\arguments{
\item{file_name_in}{path to the odm file}
}
\description{
Read the dimension of the ODM into memory
}
