% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{read_feature_ids}
\alias{read_feature_ids}
\title{Read the feature IDs of an ODM into memory}
\usage{
read_feature_ids(file_name_in, n_features)
}
\arguments{
\item{file_name_in}{path to the odm file}

\item{n_features}{the number of features in the dataset}
}
\description{
Read the feature IDs of an ODM into memory
}
