% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/high_level_initialize_helper.R
\name{read_given_column_of_tsv}
\alias{read_given_column_of_tsv}
\title{Read given column of tsv}
\usage{
read_given_column_of_tsv(col_idx, n_cols, tsv_file)
}
\arguments{
\item{col_idx}{index of column to read}

\item{n_cols}{number of columns in file}

\item{tsv_file}{file path to .tsv file}
}
\value{
contents of the specified column in vector form
}
\description{
Read given column of tsv
}
