% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{read_integer_vector}
\alias{read_integer_vector}
\title{Read the dimension of the ODM into memory}
\usage{
read_integer_vector(file_name_in, dataset_name, length)
}
\arguments{
\item{file_name_in}{path to the odm file}

\item{dataset_name}{name of the dataset within the file}

\item{length}{dimension of the integer vector to read}
}
\description{
Read the dimension of the ODM into memory
}
