% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{read_integer_vector_hdf5}
\alias{read_integer_vector_hdf5}
\title{read intege vector from hdf5}
\usage{
read_integer_vector_hdf5(file_name_in, dataset_name_in, data_len)
}
\arguments{
\item{file_name_in}{name of h5 file}

\item{dataset_name_in}{name of dataset to read}

\item{data_len}{length of integer vector}
}
\description{
read intege vector from hdf5
}
