% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{read_row_ptr}
\alias{read_row_ptr}
\title{Read the row pointer into memory}
\usage{
read_row_ptr(file_name_in, n_features)
}
\arguments{
\item{file_name_in}{path to the odm file}

\item{n_features}{the number of features in the dataset}
}
\description{
Read the row pointer into memory
}
