% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/low_level_indexing.R
\name{return_spMatrix_contig_index}
\alias{return_spMatrix_contig_index}
\title{Return sparse matrix from contiguous index}
\usage{
return_spMatrix_contig_index(
  h5_file,
  contiguous_idx_range,
  index_on_cell,
  logical_mat,
  underlying_dimension
)
}
\arguments{
\item{h5_file}{the file path to the h5 file}

\item{contiguous_idx_range}{a length-two INTEGER vector containing the starting end ending index (inclusive)}

\item{index_on_cell}{(boolean) index on cell (TRUE) or feature (FALSE)}

\item{logical_mat}{(boolean)}

\item{underlying_dimension}{a length 2 integer vector giving the underlying dimension of the expression matrix.}
}
\value{
a Matrix object with the given contiguous_idx_range.
}
\description{
Return sparse matrix from contiguous index
}
