% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/low_level_indexing.R
\name{return_spMatrix_from_contiguous_index}
\alias{return_spMatrix_from_contiguous_index}
\title{Return sparse matrix from contiguous index}
\usage{
return_spMatrix_from_contiguous_index(h5_file, h5_group, contiguous_idx_range)
}
\arguments{
\item{h5_file}{path to h5 file on disk.}

\item{h5_group}{h5 group from which to extract data (compressed_sparse_row/ or compressed_sparse_column/)}

\item{contiguous_idx_range}{an integer vector of length 2 specifying the (inclusive) range of rows/columns to extract}
}
\value{
a list containing the 0-based pointer vector (excluding 0 itself) and the associated data
}
\description{
Takes an index range (length two integer vector), h5_file, and h5_group (column or row) and returns the associated data, indexes, and pointer.
}
