% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/low_level_initialize.R
\name{run_core_algo_step_1_mtxchunked}
\alias{run_core_algo_step_1_mtxchunked}
\title{Run mtx algo step 1 for single .mtx file in chunk mode}
\usage{
run_core_algo_step_1_mtxchunked(
  mtx_fp,
  initialize_accumulator,
  bag_of_variables,
  is_logical,
  n_lines_per_chunk,
  n_rows_to_skip,
  progress
)
}
\description{
Runs the first step of the .mtx algo.
}
