% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/low_level_initialize.R
\name{run_core_mtx_algo}
\alias{run_core_mtx_algo}
\title{Run core mtx algo}
\usage{
run_core_mtx_algo(
  h5_fp,
  mtx_fp,
  is_logical,
  covariates,
  bag_of_variables,
  n_lines_per_chunk,
  n_rows_to_skip
)
}
\arguments{
\item{h5_fp}{file path to h5 file}

\item{mtx_fp}{file path to .mtx file}

\item{is_logical}{(logical) is expression matrix logical?}

\item{covariates}{a list of two elements: the feature covariates and the cell covariates}

\item{bag_of_variables}{environment containing named variables}

\item{n_lines_per_chunk}{number of elements to load per chunk}

\item{n_rows_to_skip}{number of rows to skip in .mtx file}
}
\description{
Runs to core algorithm for mtx files. There are two steps:
(i) compute the row pointer
(ii) write CSC matrix, write CSR matrix, compute covariate matrices
}
