% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/low_level_initialize.R
\name{run_mtx_algo_step_1}
\alias{run_mtx_algo_step_1}
\title{Run mtx algo step 1}
\usage{
run_mtx_algo_step_1(
  mtx_fp,
  initialize_accumulator,
  bag_of_variables,
  is_logical,
  n_elem_per_chunk,
  n_rows_to_skip
)
}
\arguments{
\item{mtx_fp}{file path to .mtx file}

\item{initialize_accumulator}{initialize accumulator function}

\item{bag_of_variables}{the bag of variables}

\item{is_logical}{(boolean) is matrix logical}

\item{n_elem_per_chunk}{number of elements to process per chunk}

\item{n_rows_to_skip}{number of rows to skip in .mtx file}
}
\value{
list containing (i) n_features, and (ii) a list containing n_features an n_features vector for each chunk.
}
\description{
Runs the first step of the .mtx algo.
}
