% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/low_level_initialize.R
\name{run_mtx_algo_step_2}
\alias{run_mtx_algo_step_2}
\title{Run mtx algo step 2}
\usage{
run_mtx_algo_step_2(
  h5_fp,
  mtx_fp,
  is_logical,
  bag_of_variables,
  n_elem_per_chunk,
  n_rows_to_skip,
  initial_accumulators,
  terminal_functs_args
)
}
\arguments{
\item{h5_fp}{full path to the h5 file on-disk}

\item{bag_of_variables}{the bag of variables containing the variables to pass to the accumulator functions}

\item{n_elem_per_chunk}{number of elements to load per chunk}

\item{initial_accumulators}{list of starting accumulators}

\item{terminal_functs_args}{list of accumulator function names and arguments}
}
\value{
a list containing the values of the terminals
}
\description{
This function runs step 2 of the core mtx algorithm. It (a) computes the terminal symbols, (b) writes to the CSC matrix, and (b) sorts the data by feature_idx, then writes to the CSR matrix.
}
