% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/low_level_initialize.R
\name{run_subtask_2b}
\alias{run_subtask_2b}
\title{Run subtask 2b}
\usage{
run_subtask_2b(x, pos, h5_fp, is_logical)
}
\arguments{
\item{x}{a data.table}

\item{pos}{the starting row in the data.table; uses 1-based indexing}

\item{h5_fp}{file path to the on-disk h5 file.}

\item{is_logical}{is the mtx logical}
}
\description{
Runs subtask 2b of algorithm: Writes chunk of gene_idxs and (if integer matrix) data to CSC matrix on-disk.
}
\details{
NOTE: parts of this function should be rewritten in C++; in particular, we may need to handle pos manually, as r integers have limited size.
}
