% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/low_level_initialize.R
\name{run_subtask_2c}
\alias{run_subtask_2c}
\title{Run subtask 2c}
\usage{
run_subtask_2c(
  x,
  h5_fp,
  is_logical,
  row_ptr,
  n_nonzero_features_per_chunk,
  chunk_no,
  n_features
)
}
\arguments{
\item{x}{a data.table}

\item{h5_fp}{file path to on-disk h5 file}

\item{is_logical}{(boolean) is the matrix logical?}

\item{row_ptr}{the (accumulated) row pointer}

\item{n_nonzero_features_per_chunk}{a list of the number of nonzero features in each chunk}

\item{chunk_no}{the current chunk number}

\item{n_features}{total number of features in matrix}
}
\value{

}
\description{
Run subtask 2c
}
