% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_functions.R
\name{save_random_matrix_as_10x}
\alias{save_random_matrix_as_10x}
\title{Save random matrix as a 10X object}
\usage{
save_random_matrix_as_10x(m, data_dir, idx = NULL)
}
\arguments{
\item{m}{a sparse Matrix object}

\item{data_dir}{the directory in which to store the matrix}

\item{idx}{(optional) an index to append to the file names}
}
\value{
the file paths to the matrix.mtx, barcodes.tsv, and features.tsv files.
}
\description{
This function stores a matrix in .mtx format, along with features and barcodes .tsv files for the cells and genes.
}
