% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{setNames}
\alias{setNames}
\title{setNames}
\usage{
setNames(object, nm)
}
\arguments{
\item{object}{the object to name}

\item{nm}{the names}
}
\value{
the named object
}
\description{
My personal setNames function
}
