% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s4_classes_methods.R
\name{show,covariate_ondisc_matrix-method}
\alias{show,covariate_ondisc_matrix-method}
\title{Print basic information to console}
\usage{
\S4method{show}{covariate_ondisc_matrix}(object)
}
\arguments{
\item{objet}{a covariate_ondisc_matrix to show}
}
\description{
Print basic information to console
}
