% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s4_classes_methods.R
\name{show,multimodal_ondisc_matrix-method}
\alias{show,multimodal_ondisc_matrix-method}
\title{Print basic information to console}
\usage{
\S4method{show}{multimodal_ondisc_matrix}(object)
}
\arguments{
\item{object}{a multimodal_ondisc_matrix to show}
}
\description{
Print basic information to console
}
\examples{
#' # NOTE: You must create the RDS files "expressions.rds" and
# "perturbations.rds" to run this example. Navigate to the help file of
# "create_ondisc_matrix_from_mtx" (via ?create_ondisc_matrix_from_mtx),
# and execute both code blocks.
expression_fp <- paste0(tempdir(), "/expressions.rds")
perturbations_fp <- paste0(tempdir(), "/perturbations.rds")
if (file.exists(expression_fp) && file.exists(perturbations_fp)) {
    expressions <- readRDS(expression_fp)
    perturbations <- readRDS(expression_fp)
    crispr_experiment <- multimodal_ondisc_matrix(list(expressions = expressions,
    perturbations = perturbations))
    show(crispr_experiment)
}
}
