% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s4_classes_methods.R
\name{show,on_disc_matrix-method}
\alias{show,on_disc_matrix-method}
\title{print first few rows and columns of an on_disc_matrix to the console; also display object class}
\usage{
\S4method{show}{on_disc_matrix}(object)
}
\arguments{
\item{object}{an on_dist_matrix to show}
}
\description{
print first few rows and columns of an on_disc_matrix to the console; also display object class
}
