% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s4_classes_methods.R
\name{show,ondisc_matrix-method}
\alias{show,ondisc_matrix-method}
\title{Print basic information to console}
\usage{
\S4method{show}{ondisc_matrix}(object)
}
\arguments{
\item{object}{an on_dist_matrix to show}
}
\description{
Also display object class, number of rows, and number of columns.
}
\examples{
# NOTE: You must create the HDF5 file "expressions.h5" to run this example.
# Navigate to the help file of "create_ondisc_matrix_from_mtx"
# (via ?create_ondisc_matrix_from_mtx), and execute the code in the first code block.

h5_fp <- paste0(tempdir(), "/expressions.h5")
if (file.exists(h5_fp)) {
odm <- ondisc_matrix(h5_file = h5_fp)
show(odm)
}
}
