% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s4_get_basic_info_methods.R
\name{show}
\alias{show}
\alias{show,ondisc_matrix-method}
\alias{show,metadata_ondisc_matrix-method}
\alias{show,multimodal_ondisc_matrix-method}
\title{Print basic information to the console}
\usage{
\S4method{show}{ondisc_matrix}(object)

\S4method{show}{metadata_ondisc_matrix}(object)

\S4method{show}{multimodal_ondisc_matrix}(object)
}
\arguments{
\item{object}{an object of class ondisc_matrix, covaraite_ondisc_matrix, or multimodal_ondisc_matrix}
}
\description{
Print basic information to the console
}
\examples{
# NOTE: You must create the HDF5 file "expressions.h5" to run this example.
# Navigate to the help file of "create_ondisc_matrix_from_mtx"
# (via ?create_ondisc_matrix_from_mtx), and execute the code in the first code block.

h5_fp <- paste0(tempdir(), "/expressions.h5")
if (file.exists(h5_fp)) {
odm <- ondisc_matrix(h5_file = h5_fp)
show(odm)
}
}
