% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s4_classes_methods.R
\name{subset-multimodal-odm}
\alias{subset-multimodal-odm}
\alias{[,multimodal_ondisc_matrix,missing,missing,missing-method}
\alias{[,multimodal_ondisc_matrix,missing,ANY,missing-method}
\alias{[,multimodal_ondisc_matrix,ANY,missing,missing-method}
\alias{[,multimodal_ondisc_matrix,ANY,ANY,missing-method}
\title{Subset a \code{multimodal_ondisc_matrix} with \code{[}}
\usage{
\S4method{[}{multimodal_ondisc_matrix,missing,missing,missing}(x, i, j, drop)

\S4method{[}{multimodal_ondisc_matrix,missing,ANY,missing}(x, i, j, drop)

\S4method{[}{multimodal_ondisc_matrix,ANY,missing,missing}(x, i, j, drop)

\S4method{[}{multimodal_ondisc_matrix,ANY,ANY,missing}(x, i, j, drop)
}
\arguments{
\item{x}{A multimodal_ondisc_matrix object}

\item{i}{not used}

\item{j}{Vector (numeric, logical, or character) indicating cells to keep}

\item{drop}{not used}
}
\value{
A subset covariate_ondisc_matrix object.
}
\description{
The user can pass logical, character, of numeric vectors to \code{`[`}. Character vectors correspond to cell barcodes.
Multimodal_ondisc_matrix objects can be subset only by cell, not feature, as different modalities (in general) have different numbers of features.
}
\examples{
# NOTE: You must create the RDS files "expressions.rds" and
# "perturbations.rds" to run this example. Navigate to the help file of
# "create_ondisc_matrix_from_mtx" (via ?create_ondisc_matrix_from_mtx),
# and execute both code blocks.
expression_fp <- paste0(tempdir(), "/expressions.rds")
perturbations_fp <- paste0(tempdir(), "/perturbations.rds")
if (file.exists(expression_fp) && file.exists(perturbations_fp)) {
    expressions <- readRDS(expression_fp)
    perturbations <- readRDS(expression_fp)
    crispr_experiment <- multimodal_ondisc_matrix(list(expressions = expressions,
    perturbations = perturbations))
    # Keep all cells except 10,100, and 105.
    x <- crispr_experiment[,-c(10,100,105)]
}
}
