% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s4_classes_methods.R
\name{subset-odm}
\alias{subset-odm}
\alias{[,ondisc_matrix,missing,missing,missing-method}
\alias{[,ondisc_matrix,ANY,missing,missing-method}
\alias{[,ondisc_matrix,missing,ANY,missing-method}
\alias{[,ondisc_matrix,ANY,ANY,missing-method}
\title{Subset an \code{ondisc_matrix} with \code{[}}
\usage{
\S4method{[}{ondisc_matrix,missing,missing,missing}(x, i, j, drop)

\S4method{[}{ondisc_matrix,ANY,missing,missing}(x, i, j)

\S4method{[}{ondisc_matrix,missing,ANY,missing}(x, i, j)

\S4method{[}{ondisc_matrix,ANY,ANY,missing}(x, i, j)
}
\arguments{
\item{x}{An ondisc_matrix object}

\item{i}{Vector (numeric, logical, or character) indicating features to keep}

\item{j}{Vector (numeric, logical, or character) indicating cells to keep}

\item{drop}{not used}
}
\value{
A subset ondisc_matrix object.
}
\description{
The user can pass logical, character, of numeric vectors to \code{`[`}. Character vectors correspond to feature IDs (for rows) and cell barcodes (for columns).
}
\examples{
# NOTE: You must create the HDF5 file "expressions.h5" to run this example.
# Navigate to the help file of "create_ondisc_matrix_from_mtx"
# (via ?create_ondisc_matrix_from_mtx), and execute the code in the first code block.

h5_fp <- paste0(tempdir(), "/expressions.h5")
if (file.exists(h5_fp)) {
odm <- ondisc_matrix(h5_file = h5_fp)
# keep cells 100-110
x <- odm[,100:110]
# keep the following genes:
x <- odm[c("ENSG00000188305", "ENSG00000257284", "ENSG00000251655"),]
# keep the following cells:
x <- odm[,c("CTTAGGACACTGGCGT-1", "AAAGGATTCACATCAG-1")]
# keep all genes except ENSG00000188305 and ENSG00000257284
x <- odm[!(get_feature_ids(odm) \%in\% c("ENSG00000188305", "ENSG00000257284")),]
}
}
