% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/high_level_matrix_ops.R
\name{subset_by_feature_or_cell}
\alias{subset_by_feature_or_cell}
\title{Subset by feature or cell}
\usage{
subset_by_feature_or_cell(x, idx, subset_on_cell, subset_string_arrays = FALSE)
}
\arguments{
\item{x}{an ondisc_matrix}

\item{idx}{a numeric, character, or logical index}

\item{subset_on_cell}{(boolean) subset on cell (TRUE) or feature (FALSE)}
}
\value{
a subset ondisc_matrix
}
\description{
Subsets an ondisc_matrix by either feature or cell.
}
