% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/high_level_matrix_ops.R
\name{subset_by_gene_or_cell}
\alias{subset_by_gene_or_cell}
\title{Subset by gene or cell}
\usage{
subset_by_gene_or_cell(x, idx, subset_on_cell)
}
\arguments{
\item{x}{an on_disc_matrix}

\item{idx}{a set of integer, logical, or character indexes.}

\item{subset_on_cell}{boolean indicating whether to subset on cell (TRUE) or gene (FALSE)}
}
\value{
a subset on_disc_matrix.
}
\description{
Subsets an on_disc_matrix by either gene or cell, as specified.
}
