% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{sum_in_place}
\alias{sum_in_place}
\title{add vectors in-place}
\usage{
sum_in_place(v1, v2)
}
\arguments{
\item{v1}{the vector to be modified in-place}

\item{v2}{the vector to add element-wise to vector v1; note: v1 may be longer than v2}
}
\description{
add vectors in-place
}
