% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/high_level_matrix_ops.R
\name{summarize_expression_matrix}
\alias{summarize_expression_matrix}
\title{Summarize expression matrix}
\usage{
summarize_expression_matrix(x, chunk_size = 4000)
}
\arguments{
\item{x}{an on_disc_matrix}

\item{chunk_size}{number of cells to process at a time}
}
\value{
a list containing the cell-specific and gene-specific covariate matrices
}
\description{
Takes an on_disc_matrix that contains gene-by-cell expressions (x), and outputs the cell-specific and gene-specific covariate matrices.
}
