% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/covariate_computation_functs.R
\name{symbols_enum}
\alias{symbols_enum}
\title{Covariate Enum
Function to return a simple "enum" of all symbols (terminal and nonterminal) used in the grammar; helps to reduce number of string literals in the code.}
\usage{
symbols_enum()
}
\value{
an environment
}
\description{
Covariate Enum
Function to return a simple "enum" of all symbols (terminal and nonterminal) used in the grammar; helps to reduce number of string literals in the code.
}
