% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crispr_screen.R
\name{thin_multimodal_odm}
\alias{thin_multimodal_odm}
\title{Thin multimodal ODM}
\usage{
thin_multimodal_odm(multimodal_odm, grna_modality_name, gene_modality_name)
}
\arguments{
\item{multimodal_odm}{a multimodal ondisc matrix}

\item{grna_modality_name}{the name of the grna modality}

\item{gene_modality_name}{the name of the gene modality}
}
\value{
a "thinned" multimodal ondisc matrix
}
\description{
"Thins" a multimodal ODM representing a single-cell CRISPR screen experiment by removing superflous and duplicate information.
}
