% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/low_level_initialize_transpose.R
\name{transpose_on_disc_csc_matrix}
\alias{transpose_on_disc_csc_matrix}
\title{Transpose on-disc CSC matrix.}
\usage{
transpose_on_disc_csc_matrix(h5_file, cell_chunk_size = 5000)
}
\arguments{
\item{h5_file}{an h5 file on disk.}

\item{cell_chunk_size}{Load chunks of the data containing this many cells.}
}
\value{
invisible
}
\description{
Converts a CSC matrix to a CSR matrix. The CSC matrix is assumed to be stored in an h5 file with fields compressed_sparse_column/p, compressed_sparse_column/data, and compressed_sparse_row/p. The function writes compressed_sparse_row/data. Indexing note: The row and column pointers are assumed to be stored using 1-based indexing. The gene indexes, as stored in compressed_sparse_column/data, use 0-based indexing. The cell indexes to be saved to compressed_sparse_row/data will be saved using 0-based indexing.
}
