% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/high_level_initialize_helper.R
\name{verify_fp}
\alias{verify_fp}
\title{Verify fp}
\usage{
verify_fp(fp)
}
\arguments{
\item{fp}{a file path}
}
\value{

}
\description{
Verifies that a file path to be used as the storage location for an ondisc matrix is acceptable.
}
\details{
There are three cases:
\itemize{
\item case 1. The directory does not exist; return TRUE, and as a side-effect, create the empty directory.
\item case 2. The directory does exist and is empty; return TRUE.
\item case 3. The directory does exist and is nonempty; return FALSE.
}
}
