% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{write_data_h5}
\alias{write_data_h5}
\title{write data h5}
\usage{
write_data_h5(file_name_in, dataset_name_in, buffer, start_pos)
}
\arguments{
\item{file_name_in}{name of h5 file}

\item{dataset_name_in}{name of the dataset within the h5 file}

\item{buffer}{the vector of integers}

\item{start_pos}{the position in the dataset in which to place the buffer}
}
\description{
write data h5
}
