% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_creation_functs.R
\name{write_example_cellranger_dataset}
\alias{write_example_cellranger_dataset}
\title{Write example cellranger dataset}
\usage{
write_example_cellranger_dataset(
  n_features,
  n_cells,
  n_batch,
  modalities,
  dir_to_write,
  p_zero = NULL,
  p_set_col_zero = NULL,
  p_set_row_zero = NULL
)
}
\arguments{
\item{n_cells}{number of cells}

\item{n_batch}{number of batches}

\item{modalities}{string indicating the modalities to create}

\item{dir_to_write}{the directory in which to write the files}

\item{n_feature_vect}{number of features per matrix}
}
\description{
Write example cellranger dataset
}
\examples{
n_features <- c(1000, 40, 400)
modalities <- c("gene", "protein", "grna")
n_cells <- 10000
n_batch <- 2
dir_to_write <- tempdir()
write_example_cellranger_dataset(n_features, n_cells, n_batch, modalities, dir_to_write)
}
