% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s4_classes_methods.R
\name{apply,on_disc_matrix-method}
\alias{apply,on_disc_matrix-method}
\title{Apply an arbitrary function to all rows or columns}
\usage{
\S4method{apply}{on_disc_matrix}(X, MARGIN, FUN, chunk_size = 4000)
}
\arguments{
\item{X}{an on_disc_matrix}

\item{MARGIN}{apply to rows (1) or columns (2)}

\item{FUN}{a function to apply}

\item{chunk_size}{number of rows or columns to load at a time; default 4000}
}
\description{
Apply an arbitrary function to all rows or columns
}
\examples{
# NOTE: You must create the HDF5 file "example.h5" to run this example.
# Navigate to the help file of "create_on_disc_matrix_from_10x_mtx"
# (via ?create_on_disc_matrix_from_10x_mtx), and execute the code in the example.
odm_fp <- system.file("extdata", "example.h5", package = "ondisc")
if (odm_fp != "") { # if required file exists, ...
odm <- on_disc_matrix(h5_file = odm_fp)
gene_means_and_sds <- apply(X = odm,
MARGIN = 1,
FUN = function(r) c(mean = mean(r), sd = sd(r)),
chunk_size = 500)
}
}
