% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ooc_pca.R
\name{compute_ooc_pca}
\alias{compute_ooc_pca}
\title{Compute out-of-core PCA on an \code{ondisc} matrix}
\usage{
compute_ooc_pca(
  covariate_odm_norm,
  fraction_variability_explained = 0.8,
  max_num_eigenvector = 30
)
}
\arguments{
\item{covariate_odm_norm}{a \code{covariate_ondisc_matrix} object}

\item{fraction_variability_explained}{the fraction variability explained}

\item{max_num_eigenvector}{the maximal number of principle components}
}
\description{
Compute out-of-core PCA on an \code{ondisc} matrix
}
\examples{
library(magrittr)
odm_fp <- system.file("extdata", "odm/gene/matrix.odm", package = "ondiscdata")
metadata_fp <- system.file("extdata", "odm/gene/metadata.rds", package = "ondiscdata")
covariate_odm <- read_odm(odm_fp, metadata_fp)
highly_expressed_feats <- sample(get_highly_expressed_features(covariate_odm, 0.01))
covariate_odm <- covariate_odm[highly_expressed_feats,]
highly_variable_feats <- get_highly_variable_features(covariate_odm, 2000)
covariate_odm <- covariate_odm[highly_variable_feats,]
covariate_odm <- covariate_odm \%>\%
mutate_cell_covariates(lg_n_umis = log(n_umis), lg_n_nonzero = log(n_nonzero))
# regress on the default covariates, i.e. p_mito and lg_n_nonzero
covariate_odm_norm <- normalize_by_regression(covariate_odm)
# run PCA covariate_odm_norm
compute_ooc_pca(covariate_odm_norm, fraction_variability_explained = 0.1)
}
