% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/high_level_matrix_ops.R
\name{convert_to_seurat_object}
\alias{convert_to_seurat_object}
\title{Convert an \code{ondisc_matrix} to a Seurat object}
\usage{
convert_to_seurat_object(
  x,
  project = "CreateSeuratObject",
  assay = "RNA",
  cell_covariate_matrix = NULL
)
}
\arguments{
\item{x}{an ondisc_matrix}

\item{project}{(default "CreateSeuratObject") project name of the Seurat object}

\item{assay}{(default "RNA") name of the initial assay}

\item{cell_covariate_matrix}{(default none) cell-level covariate matrix}
}
\value{
a Seurat object
}
\description{
Converts an ondisc_matrix to a Suerat object.
}
\examples{
# NOTE: You must create the HDF5 file "expressions.h5" to run this example.
# Navigate to the help file of "create_ondisc_matrix_from_mtx"
# (via ?create_ondisc_matrix_from_mtx), and execute the code in the first code block.
h5_fp <- paste0(tempdir(), "/expressions.h5")
if (file.exists(h5_fp)) {
odm <- ondisc_matrix(h5_file = h5_fp)
seurat_object <- convert_to_seurat_object(odm[1:100,1:100])
}
}
