% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s4_classes_methods.R
\docType{class}
\name{covariate_ondisc_matrix-class}
\alias{covariate_ondisc_matrix-class}
\alias{covariate_ondisc_matrix}
\title{covariate_ondisc_matrix class constructor}
\value{
a covariate_ondisc_matrix object
}
\description{
A covariate_ondisc_matrix (short for "covariate ondisc_matrix") is an object that stores and ondisc_matrix, along with its cell-specific and feature-specific covariate matrices.
}
\section{Slots}{

\describe{
\item{\code{ondisc_matrix}}{an ondisc_matrix.}

\item{\code{cell_covariates}}{a data frame on cell covariates.}

\item{\code{feature_covariates}}{a data frame of feature covariates.}
}}

\examples{
# NOTE: You must create the HDF5 file "expressions.h5" and the RDS file
# "expressions.rds" to run this example. Navigate to the help file of
# "create_ondisc_matrix_from_mtx" (via ?create_ondisc_matrix_from_mtx),
# and execute the code in the first code block.
covariates_fp <- paste0(tempdir(), "/expressions.rds")
h5_fp <-  paste0(tempdir(), "/expressions.h5")
if (file.exists(covariates_fp) && file.exists(h5_fp)) {
covariate_odm <- readRDS(covariates_fp)
cell_covariate_matrix <- covariate_odm@cell_covariates
feature_covariate_matrix <- covariate_odm@feature_covariates
covariate_odm_copy <- covariate_ondisc_matrix(
ondisc_matrix = ondisc_matrix(h5_file = h5_fp),
cell_covariates = cell_covariate_matrix,
feature_covariates = feature_covariate_matrix)
}
}
