% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/low_level_initialize_transpose.R
\name{create_h5_file_on_disk}
\alias{create_h5_file_on_disk}
\title{Create h5 file on disk}
\usage{
create_h5_file_on_disk(
  on_disc_dir,
  n_genes,
  n_cells,
  n_data_points,
  cell_barcodes,
  gene_ids,
  gene_names
)
}
\arguments{
\item{on_disc_dir}{directory in which to store the on_disc object}

\item{n_genes}{number of genes in dataset}

\item{n_cells}{number of cells in dataset}

\item{n_data_points}{total number of datapoints}

\item{cell_barcodes}{cell barcodes}

\item{gene_ids}{the gene ids}

\item{gene_names}{the gene names}
}
\value{
location of h5 file on disk.
}
\description{
Create h5 file on disk
}
