% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_from_cellranger.R
\name{create_odm_from_cellranger}
\alias{create_odm_from_cellranger}
\title{Create ODM from cellranger}
\usage{
create_odm_from_cellranger(
  directories_to_load,
  directory_to_write,
  write_cellwise_covariates = TRUE,
  chunk_size = 1000L,
  compression_level = 3L
)
}
\arguments{
\item{directories_to_load}{directories containing the expression data}

\item{directory_to_write}{directory in which to write the initialized \code{.odm} files and \code{cellwise_covariates.rds} file}

\item{write_cellwise_covariates}{boolean indicating whether to write the cellwise covariates to disk in \code{.rds} format}

\item{chunk_size}{chunk size to use in the backing HDF5 file}

\item{compression_level}{compression level to use in the backing HDF5 file}
}
\value{
an representing the data
}
\description{
Create ODM from cellranger
}
\examples{
\dontrun{
base_dir <- "/Users/timbarry/research_offsite/external/replogle-2022/raw/rd7/rpe1_other"
directories_to_load <- list.files(base_dir, full.names = TRUE)[1:3]
directory_to_write <- "/Users/timbarry/research_offsite/external/replogle-2022/processed/rd7_small/"
output <- create_odm_from_cellranger(directories_to_load, directory_to_write)
}
}
