% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_from_r_matrices.R
\name{create_odm_from_r_matrix}
\alias{create_odm_from_r_matrix}
\title{Create ODM from R matrix}
\usage{
create_odm_from_r_matrix(r_matrix)
}
\arguments{
\item{r_matrix}{an R matrix. The R matrix can be a standard R matrix or an object of class \code{dgTMatrix}, \code{dgCMatrix}, or \code{dgRMatrix}.}
}
\value{
a list containing the R matrix and a data frame containing the covariates
}
\description{
Creates an ODM from an R matrix.
}
