% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_data_from_r_matrices.R
\name{create_odm_from_r_matrix}
\alias{create_odm_from_r_matrix}
\title{Create ODM from R matrix}
\usage{
create_odm_from_r_matrix(
  mat,
  file_to_write,
  chunk_size = 1000L,
  compression_level = 3L
)
}
\arguments{
\item{mat}{an R matrix of class \code{"matrix"}, \code{"dgCMatrix"}, \code{"dgRMatrix"}, or \code{"dgTMatrix"}.}

\item{file_to_write}{a fully-qualified file path specifying the location in which to write the backing \code{.odm} file.}

\item{chunk_size}{(optional; default \code{1000L}) a positive integer specifying the chunk size to use to store the data in the backing HDF5 file.}

\item{compression_level}{(optional; default \code{3L}) an integer in the inveral [0, 9] specifying the compression level to use to store the data in the backing HDF5 file.}
}
\value{
an \code{odm} object
}
\description{
\code{create_odm_from_r_matrix()} converts an R matrix (stored in standard dense format or sparse format) into an \code{odm} object.
}
\examples{
library(sceptredata)
data(lowmoi_example_data)
gene_matrix <- lowmoi_example_data$response_matrix
file_to_write <- paste0(tempdir(), "gene.odm")
odm_object <- create_odm_from_r_matrix(
  mat = gene_matrix,
  file_to_write = file_to_write
)
}
