% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_data_from_r_matrices.R
\name{create_odm_from_r_matrix}
\alias{create_odm_from_r_matrix}
\title{Create ODM from R matrix}
\usage{
create_odm_from_r_matrix(
  mat,
  file_to_write,
  chunk_size = 1000L,
  compression_level = 3L,
  integer_id = 0L
)
}
\arguments{
\item{mat}{an R matrix of class "matrix", "dgCMatrix", "dgRMatrix", or "dgTMatrix"}

\item{file_to_write}{file path specifying the location in which to write the backing \code{.odm} file}

\item{chunk_size}{chunk size to use in the backing HDF5 file}

\item{compression_level}{compression level to use in the backing HDF5 file}

\item{integer_id}{integer ID to write to the backing \code{.odm} file; users need not specify this argument.}
}
\value{
an odm object representing the data
}
\description{
\code{create_odm_from_r_matrix()} creates an \code{odm} object from an input R matrix.
}
\note{
Users do not need to specify the \code{integer_id} argument; this argument is used by functions within the \code{sceptre} package that call \code{create_odm_from_r_matrix()}.
}
\examples{
library(Matrix)
set.seed(4)
n_row <- 500
n_col <- 800
v <- rpois(n = n_row * n_col, lambda = 1)
v[rbinom(n = n_row * n_col, size = 1, prob = 0.8) == 1] <- 0L
mat <- matrix(data = v, nrow = n_row, ncol = n_col)
rownames(mat) <- paste0("gene_", seq_len(nrow(mat)))
file_to_write <- paste0(tempdir(), "/gene.odm")
odm <- create_odm_from_r_matrix(mat, file_to_write)
}
