% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/high_level_initialize.R
\name{create_on_disc_matrix_from_10x_mtx}
\alias{create_on_disc_matrix_from_10x_mtx}
\title{Create on_disc_matrix from 10x .mtx file
Creates an on_disc_matrix from a 10x .mtx file.}
\usage{
create_on_disc_matrix_from_10x_mtx(
  mtx_fp,
  barcode_fp,
  features_fp,
  on_disc_dir
)
}
\arguments{
\item{mtx_fp}{file path to the .mtx file}

\item{barcode_fp}{file path to the barcode.tsv file.}

\item{features_fp}{file path to the features.tsv file.}

\item{on_disc_dir}{directory in which to store the initialized HDF5 object}
}
\value{
an on_disc_matrix object
}
\description{
Create on_disc_matrix from 10x .mtx file
Creates an on_disc_matrix from a 10x .mtx file.
}
\examples{
mtx_fp <- system.file("extdata", "matrix.mtx", package = "ondisc")
barcode_fp <- system.file("extdata", "barcodes.tsv", package = "ondisc")
features_fp <- system.file("extdata", "features.tsv", package = "ondisc")
on_disc_dir <- system.file("extdata", package = "ondisc")
if (system.file("extdata", "on_disc_matrix_1.h5", package = "ondisc") == "") {
# if on_disc_matrix does not exist, create it.
exp_mat <- create_on_disc_matrix_from_10x_mtx(mtx_fp, barcode_fp, features_fp, on_disc_dir)
}
}
