% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/initialize_in_memory.R
\name{create_ondisc_matrix_from_R_matrix}
\alias{create_ondisc_matrix_from_R_matrix}
\title{Create an \code{ondisc} matrix from an R matrix.}
\usage{
create_ondisc_matrix_from_R_matrix(
  r_matrix,
  barcodes,
  features_df,
  odm_fp,
  metadata_fp = NULL,
  gene_access_only = FALSE
)
}
\arguments{
\item{r_matrix}{an R matrix. The matrix can be represented as a standard (dense) R matrix or a sparse matrix of class "dgTMatrix". Integer and logical (i.e., boolean) matrices are allowed.}

\item{barcodes}{a character vector giving the cell barcodes.}

\item{features_df}{a data frame giving the names of the features. The first column (required) contains the feature IDs (e.g., ENSG00000186092), and the second column (optional) contains the human-readable feature names (e.g., OR4F5). Subsequent columns are discarded. Gene names starting with "MT-" are assumed to be mitochondrial genes and will be used to compute the p_mito covariate.}

\item{odm_fp}{location to write the backing .odm file.}

\item{metadata_fp}{(optional; default NULL) location to write the metadata .RDS file. By default, a file called "metadata.rds" stored in the same directory as the backing .odm file.}

\item{gene_access_only}{a boolean value, TRUE if only allow gene-wise access; FALSE if allow both gene-wise and cell-wise access}
}
\value{
A \code{covariate_ondisc_matrix} object.
}
\description{
Initializes a \code{covariate_ondisc_matrix} object from an R matrix. Returns a \code{covariate_ondisc_matrix} including an \code{ondisc_matrix} alongside cell-specific and feature-specific covariate matrices.
}
\details{
This function computes the following cell-specific and feature-specific covariates:
\itemize{
\item cell-specific: (i) total number of features expressed in cell (n_nonzero_cell), (ii) total UMI count (n_umis_cell), and (iii) percentage of UMIs that map to mitochondrial genes (p_mito_cell).
\item feature-specific: (i) total number of cells in which feature is expressed (n_nonzero_feature), (ii) mean expression of feature across cells (mean_expression_feature), (iii) coefficient of variation of feature expression across cells (coef_of_variation_feature).
}
}
\examples{
# Please install the `ondiscdata` package before running the examples.
# install.packages("devtools")
# devtools::install_github("Katsevich-Lab/ondiscdata")

###########################
# EXAMPLE 1: integer counts
###########################
file_locs <- system.file("extdata/r_matrix/gene", package = "ondiscdata",
c("matrix.rds", "features.rds", "barcodes.rds"))
r_matrix <- readRDS(file = file_locs[1])
features_df <- readRDS(file = file_locs[2])
barcodes <- readRDS(file = file_locs[3])
odm_fp <- paste0(create_new_directory(), "/integer_odm")
odm_integer <- create_ondisc_matrix_from_R_matrix(r_matrix, barcodes,
features_df, odm_fp)

####################
# EXAMPLE 2: logical
####################
file_locs <- system.file("extdata/r_matrix/gRNA", package = "ondiscdata",
c("matrix.rds", "features.rds", "barcodes.rds"))
r_matrix_2 <- readRDS(file = file_locs[1])
features_df_2 <- readRDS(file = file_locs[2])
barcodes <- readRDS(file = file_locs[3])
odm_fp <- paste0(create_new_directory(), "/logical_odm")
odm_logical <- create_ondisc_matrix_from_R_matrix(r_matrix_2, barcodes,
features_df_2, odm_fp)
}
