% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_functions.R
\name{create_synthetic_data}
\alias{create_synthetic_data}
\title{Create synthetic data}
\usage{
create_synthetic_data(
  n_datasets,
  simulated_data_dir,
  n_row = NULL,
  n_col = NULL,
  seed = NULL,
  idx_start = 1L
)
}
\arguments{
\item{n_datasets}{number of datasets to generate}

\item{simulated_data_dir}{directory in which to store the generated datasets}

\item{n_row}{number of rows in datasets (by default random)}

\item{n_col}{number of columns in datasets (by default random)}

\item{seed}{(optional) seed to set}

\item{idx_start}{index at which to start (default 1)}
}
\description{
Generate synthetic datasets (consisting of a matrix.mtx file, a barcodes.tsv file, and a features.tsv file) and store these datasets in simulated_data_dir, with indices appended to the file names.
}
