% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_functions.R
\name{create_synthetic_data}
\alias{create_synthetic_data}
\title{Create synthetic data}
\usage{
create_synthetic_data(n_row, n_col, logical_mat, write_as_mtx_to_disk = TRUE)
}
\arguments{
\item{n_row}{number of rows}

\item{n_col}{number of columns}

\item{logical_mat}{boolean indicating whether a given dataset is logical (TRUE) or integer (FALSE).}

\item{write_as_mtx_to_disk}{boolean indicating whether to write the mtx file, barcodes file, and features file to disk as side-effect.}
}
\value{
a list:
(i) R matrix
(ii) in-memory cell barcodes
(iii) in-memory features df
(iv) number of nonzero entries in expression matrix
(v) file path to the .mtx file
(vi) file path to the barcodes.tsv file
(vii) file path to the features.tsv file
}
\description{
Creates synthetic data for testing purposes. Optionally writes the data to disk.
}
\examples{
n_row <- 1000
n_col <- 500
logical_mat <- FALSE
synth_data <- create_synthetic_data(n_row, n_col, logical_mat)
# side-effect: creates .mtx file, barcodes.tsv file, and features.tsv file on disk.
}
