% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s4_classes_methods.R
\name{dim,ondisc_matrix-method}
\alias{dim,ondisc_matrix-method}
\title{Return dimension}
\usage{
\S4method{dim}{ondisc_matrix}(x)
}
\arguments{
\item{x}{an ondisc_matrix}
}
\value{
an integer vector containing the dimension of the matrix
}
\description{
Return dimension of ondisc_matrix.
}
\examples{
# NOTE: You must create the HDF5 file "expressions.h5" to run this example.
# Navigate to the help file of "create_ondisc_matrix_from_mtx"
# (via ?create_ondisc_matrix_from_mtx), and execute the code in the first code block.
h5_fp <- paste0(tempdir(), "/expressions.h5")
if (file.exists(h5_fp)) {
odm <- ondisc_matrix(h5_file = h5_fp)
dim(odm)
}
}
