% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s4_classes_methods.R
\name{dim,ondisc_matrix-method}
\alias{dim,ondisc_matrix-method}
\title{index_vec class
A union of the numeric, logical, and character classes. Used for indexing of on_disc_matrices. Idea borrowed from Matrix package.
Return dimension}
\usage{
\S4method{dim}{ondisc_matrix}(x)
}
\arguments{
\item{x}{an ondisc_matrix}
}
\value{
an integer vector containing the dimension of the matrix
}
\description{
Return dimension of ondisc_matrix.
}
\examples{
# NOTE: You must create the HDF5 file "example.h5" to run this example.
# Navigate to the help file of "create_on_disc_matrix_from_10x_mtx"
# (via ?create_on_disc_matrix_from_10x_mtx), and execute the code in the example.
odm_fp <- system.file("extdata", "example.h5", package = "ondisc")
if (odm_fp != "") { # if required file exists, ...
odm <- ondisc_matrix(h5_file = odm_fp)
dim(odm)
}
}
