% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s4_get_basic_info_methods.R
\name{dim}
\alias{dim}
\alias{dim,ondisc_matrix-method}
\alias{dim,metadata_ondisc_matrix-method}
\alias{dim,multimodal_ondisc_matrix-method}
\alias{ncol}
\alias{nrow}
\alias{ncol,multimodal_ondisc_matrix-method}
\alias{nrow,multimodal_ondisc_matrix-method}
\title{Get dimension}
\usage{
\S4method{dim}{ondisc_matrix}(x)

\S4method{dim}{metadata_ondisc_matrix}(x)

\S4method{dim}{multimodal_ondisc_matrix}(x)

ncol(x)

nrow(x)

\S4method{ncol}{multimodal_ondisc_matrix}(x)

\S4method{nrow}{multimodal_ondisc_matrix}(x)
}
\arguments{
\item{x}{an \code{ondisc_matrix}, \code{metadata_ondisc_matrix}, or \code{multimodal_ondisc_matrix}.}
}
\value{
If \code{x} is an \code{ondisc_matrix} or \code{metadata_ondisc_matrix}, length-two integer vector containing
the dimension of \code{x}; if \code{x} is a \code{multimodal_ondisc_matrix}, a list of integer vectors containing the dimensions
of the constituent modalities of \code{x}.
}
\description{
Return the dimension of an \code{ondisc_matrix}, \code{metadata_ondisc_matrix}, or \code{multimodal_ondisc_matrix}.
}
