% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s4_classes_methods.R
\name{dim}
\alias{dim}
\alias{dim,ondisc_matrix-method}
\alias{dim,metadata_ondisc_matrix-method}
\alias{dim,multimodal_ondisc_matrix-method}
\title{Return dimension}
\usage{
\S4method{dim}{ondisc_matrix}(x)

\S4method{dim}{metadata_ondisc_matrix}(x)

\S4method{dim}{multimodal_ondisc_matrix}(x)
}
\arguments{
\item{x}{an ondisc_matrix, metadata_ondisc_matrix, or multimodal_ondisc_matrix}
}
\value{
If x is an ondisc_matrix or metadata_ondisc_matrix, an integer vector containing the dimension of the matrix; if
x is a multimodal_ondisc_matrix, a list of integer vectors containing the dimension of the constituent matrices.
}
\description{
Return dimension of ondisc_matrix, metadata_ondisc_matrix, or multimodal_ondisc_matrix.
}
\examples{
# NOTE: You must create the HDF5 file "expressions.h5" to run this example.
# Navigate to the help file of "create_ondisc_matrix_from_mtx"
# (via ?create_ondisc_matrix_from_mtx), and execute the code in the first code block.
h5_fp <- paste0(tempdir(), "/expressions.h5")
if (file.exists(h5_fp)) {
odm <- ondisc_matrix(h5_file = h5_fp)
dim(odm)
}
}
