% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s4_subset_extract.R
\name{extract-odm}
\alias{extract-odm}
\alias{[[,ondisc_matrix,missing,missing-method}
\alias{[[,ondisc_matrix,ANY,missing-method}
\alias{[[,ondisc_matrix,missing,ANY-method}
\alias{[[,ondisc_matrix,ANY,ANY-method}
\alias{[[,covariate_ondisc_matrix,ANY,ANY-method}
\alias{[[,covariate_ondisc_matrix,missing,ANY-method}
\alias{[[,covariate_ondisc_matrix,ANY,missing-method}
\alias{[[,covariate_ondisc_matrix,missing,missing-method}
\alias{[[,multimodal_ondisc_matrix,ANY,ANY-method}
\title{Pull a submatrix into memory using the \code{[[} operator.}
\usage{
\S4method{[[}{ondisc_matrix,missing,missing}(x, i, j)

\S4method{[[}{ondisc_matrix,ANY,missing}(x, i, j)

\S4method{[[}{ondisc_matrix,missing,ANY}(x, i, j)

\S4method{[[}{ondisc_matrix,ANY,ANY}(x, i, j)

\S4method{[[}{covariate_ondisc_matrix,ANY,ANY}(x, i, j)

\S4method{[[}{covariate_ondisc_matrix,missing,ANY}(x, i, j)

\S4method{[[}{covariate_ondisc_matrix,ANY,missing}(x, i, j)

\S4method{[[}{covariate_ondisc_matrix,missing,missing}(x, i, j)

\S4method{[[}{multimodal_ondisc_matrix,ANY,ANY}(x, i, j)
}
\arguments{
\item{x}{an \code{ondisc_matrix} object.}

\item{i}{a vector (numeric, logical, or character) indicating features to pull into memory.}

\item{j}{a vector (numeric, logical, or character) indicating cells to pull into memory.}
}
\value{
a matrix (as implemented by the Matrix package).
}
\description{
Apply the \code{[[} operator to an \code{ondisc_matrix} or \code{covariate_ondisc_matrix} to pull a submatrix into memory. You can pass logical, character,
or numeric vectors to \code{[[}; character vectors are assumed to refer to feature IDs (for rows) and cell barcodes
(for columns).
}
\details{
One can remember the difference between \code{[} and \code{[[} by recalling R lists: \code{[} is used to subset a list, and
\code{[[} is used to access elements stored \emph{inside} a list. Similarly, \code{[} is used to subset an \code{ondisc_matrix}, and
\code{[[} is used to access a submatrix usable within R.
}
\examples{
# Install the `ondiscdata` package to run the examples.
# devtools::install_github("timothy-barry/ondiscdata")

odm_fp <- system.file("extdata", "odm/gene/matrix.odm", package = "ondiscdata")
metadata_fp <- system.file("extdata", "odm/gene/metadata.rds", package = "ondiscdata")
odm <- read_odm(odm_fp, metadata_fp)

# pull a single feature
gene_id <- sample(get_feature_ids(odm), 1)
single_feature <- odm[[gene_id,]]
# pull a single cell
cell_barcode <- sample(get_cell_barcodes(odm), 1)
single_cell <- odm[[,cell_barcode]]
# numeric vector
mtx_numeric <- odm[[1:10, 1:10]]
# logical vector
set.seed(1)
nrow_logical <- sample(x = c(TRUE, FALSE), size = nrow(odm), replace = TRUE)
ncol_logical <- sample(x = c(TRUE, FALSE), size = ncol(odm), replace = TRUE)
mtx_logical <- odm[[nrow_logical, ncol_logical]]
# character vectors for feature IDs and cell barcodes
feature_ids_subset <- get_feature_ids(odm)[1:10]
cell_barcodes_subset <- get_cell_barcodes(odm)[1:10]
mtx_char <- odm[[feature_ids_subset, cell_barcodes_subset]]
}
