% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s4_classes_methods.R
\name{extract-odm}
\alias{extract-odm}
\alias{[[,on_disc_matrix,missing,missing-method}
\alias{[[,on_disc_matrix,index_vec,missing-method}
\alias{[[,on_disc_matrix,missing,index_vec-method}
\alias{[[,on_disc_matrix,index_vec,index_vec-method}
\title{Extract an on_disc_matrix
Pull a submatrix of an \linkS4class{on_disc_matrix} into memory using the \code{`[[`} operator.}
\usage{
\S4method{[[}{on_disc_matrix,missing,missing}(x, i, j)

\S4method{[[}{on_disc_matrix,index_vec,missing}(x, i, j)

\S4method{[[}{on_disc_matrix,missing,index_vec}(x, i, j)

\S4method{[[}{on_disc_matrix,index_vec,index_vec}(x, i, j)
}
\arguments{
\item{x}{An \linkS4class{on_disc_matrix} object (Do not pass as an argument to \code{`[[`}).}

\item{i}{Vector (numeric, logical, or character) indicating genes to keep.}

\item{j}{Vector (numeric, logical, or character) indicating cells to keep.}
}
\value{
A matrix object (of class Matrix)
}
\description{
Extract an on_disc_matrix
Pull a submatrix of an \linkS4class{on_disc_matrix} into memory using the \code{`[[`} operator.
}
\examples{
# exp_mat_loc <- system.file("extdata", "on_disc_matrix_1.h5", package = "ondisc")
# if (exp_mat_loc != "") {
# x <- on_disc_matrix(h5_file = exp_mat_loc) # initialze an on_disc_matrix
# x_mem <- x[[1:10,1:10]] # subset first 10 genes and cells
# }
}
