% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s4_classes_methods.R
\name{extract-odm}
\alias{extract-odm}
\alias{[[,ondisc_matrix,missing,missing-method}
\alias{[[,ondisc_matrix,ANY,missing-method}
\alias{[[,ondisc_matrix,missing,ANY-method}
\alias{[[,ondisc_matrix,ANY,ANY-method}
\title{Pull a submatrix into memory with \code{[[}}
\usage{
\S4method{[[}{ondisc_matrix,missing,missing}(x, i, j)

\S4method{[[}{ondisc_matrix,ANY,missing}(x, i, j)

\S4method{[[}{ondisc_matrix,missing,ANY}(x, i, j)

\S4method{[[}{ondisc_matrix,ANY,ANY}(x, i, j)
}
\arguments{
\item{x}{An ondisc_matrix object}

\item{i}{Vector (numeric, logical, or character) indicating features to keep}

\item{j}{Vector (numeric, logical, or character) indicating cells to keep}
}
\value{
A matrix (of class Matrix)
}
\description{
The user can pass logical, character, of numeric vectors to \code{`[[`}. Character vectors correspond to feature IDs (for rows) and cell barcodes (for columns).
}
\examples{
# NOTE: You must create the HDF5 file "expressions.h5" to run this example.
# Navigate to the help file of "create_ondisc_matrix_from_mtx"
# (via ?create_ondisc_matrix_from_mtx), and execute the code in the first code block.

h5_fp <- paste0(tempdir(), "/expressions.h5")
if (file.exists(h5_fp)) {
odm <- ondisc_matrix(h5_file = h5_fp)
# extract cells 100-110
x <- odm[[,100:110]]
# extract the following genes:
x <- odm[[c("ENSG00000188305", "ENSG00000257284", "ENSG00000251655"),]]
# extract the following cells:
x <- odm[[,c("CTTAGGACACTGGCGT-1", "AAAGGATTCACATCAG-1")]]
}
}
