% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s4_classes_methods.R
\name{extract-odm}
\alias{extract-odm}
\alias{[[,on_disc_matrix,missing,missing-method}
\alias{[[,on_disc_matrix,index_vec,missing-method}
\alias{[[,on_disc_matrix,missing,index_vec-method}
\alias{[[,on_disc_matrix,index_vec,index_vec-method}
\title{Pull a submatrix into memory with \code{[[}}
\usage{
\S4method{[[}{on_disc_matrix,missing,missing}(x, i, j)

\S4method{[[}{on_disc_matrix,index_vec,missing}(x, i, j)

\S4method{[[}{on_disc_matrix,missing,index_vec}(x, i, j)

\S4method{[[}{on_disc_matrix,index_vec,index_vec}(x, i, j)
}
\arguments{
\item{x}{An on_disc_matrix object}

\item{i}{Vector (numeric, logical, or character) indicating genes to keep}

\item{j}{Vector (numeric, logical, or character) indicating cells to keep}
}
\value{
A matrix (of class Matrix)
}
\description{
The user can pass logical, character, of numeric vectors to \code{`[[`}. Character vectors correspond to gene IDs (for rows) and cell barcodes (for columns).
}
\examples{
# NOTE: You must create the HDF5 file "example.h5" to run this example.
# Navigate to the help file of "create_on_disc_matrix_from_10x_mtx"
# (via ?create_on_disc_matrix_from_10x_mtx), and execute the code in the example.
odm_fp <- system.file("extdata", "example.h5", package = "ondisc")
if (odm_fp != "") { # if required file exists, ...
odm <- on_disc_matrix(h5_file = odm_fp)
# keep cells 100-110
x <- odm[[,100:110]]
# keep genes ENSG00000260528, ENSG00000258908, and ENSG00000081913
x <- odm[[c("ENSG00000260528", "ENSG00000258908", "ENSG00000081913"),]]
# keep cells ACAGCCGCAGAAACCG and CTACTATAGTGTACCT
x <- odm[[,c("ACAGCCGCAGAAACCG", "CTACTATAGTGTACCT")]]
}
}
