% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/high_level_matrix_ops.R
\name{extract_matrix}
\alias{extract_matrix}
\title{Extract matrix}
\usage{
extract_matrix(x)
}
\arguments{
\item{x}{an ondisc_matrix.}
}
\value{
an in-memory version of x in the form of a Matrix object.
}
\description{
Extract matrix
}
