% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/high_level_matrix_ops.R
\name{get_cell_barcodes}
\alias{get_cell_barcodes}
\alias{get_feature_ids}
\alias{get_feature_names}
\title{Get feature names, feature IDs, and cell barcodes}
\usage{
get_cell_barcodes(x)

get_feature_ids(x)

get_feature_names(x)
}
\arguments{
\item{x}{An ondisc_matrix.}
}
\value{
The cell barcodes of this on-disc matrix.

The cell feature ids of this on-disc matrix.

The feature names of this on-disc matrix.
}
\description{
Functions to extract cell barcodes, feature ids, and feature names from an ondisc_matrix.
}
\examples{
# NOTE: You must create the HDF5 file "example.h5" to run this example.
# Navigate to the help file of "create_ondisc_matrix_from_10x_mtx"
# (via ?create_ondisc_matrix_from_10x_mtx), and execute the code in the example.
odm_fp <- system.file("extdata", "example.h5", package = "ondisc")
if (odm_fp != "") { # if required file exists, ...
odm <- ondisc_matrix(h5_file = odm_fp)
barcodes <- get_cell_barcodes(odm)
feature_ids <- get_feature_ids(odm)
feature_names <- get_feature_names(odm)
}
}
