% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/high_level_matrix_ops.R
\name{get_cell_barcodes}
\alias{get_cell_barcodes}
\alias{get_gene_ids}
\alias{get_gene_names}
\title{Get gene names, gene IDs, and cell barcodes}
\usage{
get_cell_barcodes(x)

get_gene_ids(x)

get_gene_names(x)
}
\arguments{
\item{x}{An on_disc_matrix.}
}
\value{
The cell barcodes of this on-disc matrix.

The cell gene ids of this on-disc matrix.

The gene names of this on-disc matrix.
}
\description{
Functions to extract cell barcodes, gene ids, and gene names from an on_disc_matrix.
}
\examples{
# NOTE: You must create the HDF5 file "example.h5" to run this example.
# Navigate to the help file of "create_on_disc_matrix_from_10x_mtx"
# (via ?create_on_disc_matrix_from_10x_mtx), and execute the code in the example.
odm_fp <- system.file("extdata", "example.h5", package = "ondisc")
if (odm_fp != "") { # if required file exists, ...
odm <- on_disc_matrix(h5_file = odm_fp)
barcodes <- get_cell_barcodes(odm)
gene_ids <- get_gene_ids(odm)
gene_names <- get_gene_names(odm)
}
}
