% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s4_other_verbs.R
\name{get_modality}
\alias{get_modality}
\title{Get modality}
\usage{
get_modality(multimodal_mat, modality_name)
}
\arguments{
\item{multimodal_mat}{a multimodal_ondisc_matrix object.}

\item{modality_name}{name of modality to extract.}
}
\value{
the requested modality
}
\description{
Obtains a given modality from a multimodal_ondisc_matrix object by name.
}
\examples{
# Please install the `ondiscdata` package before running the examples.
# install.packages("devtools")
# devtools::install_github("Katsevich-Lab/ondiscdata")

odm_gene_fp <- system.file("extdata", "odm/gene/matrix.odm", package = "ondiscdata")
metadata_gene_fp <- system.file("extdata", "odm/gene/metadata.rds", package = "ondiscdata")
odm_gene <- read_odm(odm_gene_fp, metadata_gene_fp)
odm_gRNA_fp <- system.file("extdata", "odm/gRNA/matrix.odm", package = "ondiscdata")
metadata_gRNA_fp <- system.file("extdata", "odm/gRNA/metadata.rds", package = "ondiscdata")
odm_gRNA <- read_odm(odm_gRNA_fp, metadata_gRNA_fp)

odm_multi <- multimodal_ondisc_matrix(list(gene = odm_gene, gRNA = odm_gRNA))
odm_gene <- get_modality(odm_multi, "gene")
}
