% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/high_level_initialize_helper.R
\name{get_mtx_metadata}
\alias{get_mtx_metadata}
\title{Get mtx metadata}
\usage{
get_mtx_metadata(mtx_fp)
}
\arguments{
\item{mtx_fp}{path to the mtx file

If total number of data points exceeds maximum integer value in R (.Machine$integer.max), throw error.}
}
\value{
a list containing (i) n_genes, (ii) n_cells, (iii) the number of
data points (i.e., fraction of entries that are zero),
(iv) (TRUE/FALSE) matrix is logical,
(v) number of rows to skip before reading the data
}
\description{
Get mtx metadata
}
