% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/high_level_initialize_helper.R
\name{get_n_rows_with_comments_mtx}
\alias{get_n_rows_with_comments_mtx}
\title{Get n rows with comments}
\usage{
get_n_rows_with_comments_mtx(mtx_fp)
}
\arguments{
\item{mtx_fp}{a file path to an mtx file}
}
\value{
the number of rows with comments at the top of the file
}
\description{
Returns the number of rows with comments in an mtx file.
}
