% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s4_other_verbs.R
\name{get_ondisc_matrix}
\alias{get_ondisc_matrix}
\title{Get the ondisc_matrix stored in a \code{covariate_ondisc_matrix.}}
\usage{
get_ondisc_matrix(covariate_odm)
}
\arguments{
\item{covariate_odm}{a \code{covariate_ondisc_matrix} object}
}
\value{
the \code{ondisc_matrix} stored within the \code{covariate_ondisc_matrix} object
}
\description{
Get the ondisc_matrix stored in a \code{covariate_ondisc_matrix.}
}
\examples{
# Install the `ondiscdata` package to run the examples.
# devtools::install_github("timothy-barry/ondiscdata")

odm_fp <- system.file("extdata", "odm/gene/matrix.odm", package = "ondiscdata")
metadata_fp <- system.file("extdata", "odm/gene/metadata.rds", package = "ondiscdata")
cov_odm <- read_odm(odm_fp, metadata_fp)

odm_mtx <- get_ondisc_matrix(cov_odm)
}
