% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/high_level_initialize_helper.R
\name{get_string_arrays}
\alias{get_string_arrays}
\title{Get string}
\usage{
get_string_arrays(barcodes_fp, features_fp, features_metadata)
}
\arguments{
\item{barcodes_fp}{path to barcodes file(s)}

\item{features_fp}{path the features file}

\item{features_metadata}{the features_metadata list}
}
\value{
list containing (i) cell barcodes, (ii) feature IDs, and (iii) feature names (NA_character_ if absent)
}
\description{
Get string
}
