% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/high_level_matrix_ops.R
\name{get_subset_vector}
\alias{get_subset_vector}
\title{Get on_disc subset vector}
\usage{
get_subset_vector(x, cell_idx)
}
\arguments{
\item{x}{the on_disc_matrix.}

\item{cell_idx}{(boolean) TRUE = cell, FALSE = gene.}
}
\value{
the requested subset indexes
}
\description{
Returns the subset vector (either gene or cell) for an on_disc_matrix object.
}
