% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s4_get_basic_info_methods.R
\name{head,ondisc_matrix-method}
\alias{head,ondisc_matrix-method}
\title{head}
\usage{
\S4method{head}{ondisc_matrix}(x)
}
\arguments{
\item{x}{an \code{ondisc_matrix}.}
}
\value{
NULL; called for printing
}
\description{
Print the first few rows and columns of an \code{ondisc_matrix}.
}
