% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s4_get_basic_info_methods.R
\name{head}
\alias{head}
\alias{head,ondisc_matrix-method}
\alias{head,covariate_ondisc_matrix-method}
\alias{head,multimodal_ondisc_matrix-method}
\title{Head}
\usage{
\S4method{head}{ondisc_matrix}(x)

\S4method{head}{covariate_ondisc_matrix}(x)

\S4method{head}{multimodal_ondisc_matrix}(x)
}
\arguments{
\item{x}{an \code{ondisc_matrix}, \code{covariate_ondisc_matrix}, or \code{multimodal_ondisc_matrix}.}
}
\value{
NULL; called for printing
}
\description{
Print the first few rows and columns of an \code{ondisc_matrix}, \code{covariate_ondisc_matrix}, or \code{multimodal_ondisc_matrix}.
}
\examples{
# Use `ondiscdata` package for the examlpes, please install the package before running the examples
# install.packages("devtools")
# devtools::install_github("Katsevich-Lab/ondiscdata")

####################################
# EXAMPLE 1: covariate_ondisc_matrix
####################################
# Load odm from package
odm_fp <- system.file("extdata", "odm/gene/matrix.odm", package = "ondiscdata")
metadata_fp <- system.file("extdata", "odm/gene/metadata.rds", package = "ondiscdata")
odm <- read_odm(odm_fp, metadata_fp)
head(odm)

#####################################
# EXAMPLE 2: multimodal_ondisc_matrix
#####################################
# Load odm from package
odm_gene_fp <- system.file("extdata", "odm/gene/matrix.odm", package = "ondiscdata")
metadata_gene_fp <- system.file("extdata", "odm/gene/metadata.rds", package = "ondiscdata")
odm_gene <- read_odm(odm_gene_fp, metadata_gene_fp)
odm_gRNA_fp <- system.file("extdata", "odm/gRNA/matrix.odm", package = "ondiscdata")
metadata_gRNA_fp <- system.file("extdata", "odm/gRNA/metadata.rds", package = "ondiscdata")
odm_gRNA <- read_odm(odm_gRNA_fp, metadata_gRNA_fp)

odm_multi <- multimodal_ondisc_matrix(list(gene = odm_gene, gRNA = odm_gRNA))
head(odm_multi)
}
