% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s4_class.R
\name{initialize_odm_from_backing_file}
\alias{initialize_odm_from_backing_file}
\title{Initialize an ondisc matrix}
\usage{
initialize_odm_from_backing_file(odm_file)
}
\arguments{
\item{odm_file}{file path to the backing \code{.odm} file}
}
\value{
an object of class \code{odm}
}
\description{
Initializes an object of class \code{odm} from a backing \code{.odm} file.
}
\details{
This function is portable: one can create a \code{.odm} file (via \code{create_odm_from_cellranger()}), transfer the \code{.odm} file to another computer, and then load the \code{.odm} file (via \code{initialize_odm_from_backing_file()}) on the second computer.
}
\examples{
library(sceptredata)
directories_to_load <- paste0(
 system.file("extdata", package = "sceptredata"),
 "/highmoi_example/gem_group_", c(1, 2)
)
directory_to_write <- tempdir()
out_list <- create_odm_from_cellranger(
  directories_to_load = directories_to_load,
  directory_to_write = directory_to_write,
)
gene_odm <- out_list$gene
gene_odm

# delete the gene_odm object
rm(gene_odm)

# reinitialize the gene_odm object
gene_odm <- initialize_odm_from_backing_file(
  paste0(tempdir(), "/gene.odm")
)
gene_odm
}
