% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crispr_screen.R
\name{load_thresholded_and_grouped_grna}
\alias{load_thresholded_and_grouped_grna}
\title{Load thresholded and grouped grna}
\usage{
load_thresholded_and_grouped_grna(
  covariate_odm,
  grna_group,
  threshold = 3,
  grna_group_name = "grna_group"
)
}
\arguments{
\item{covariate_odm}{a grna-by-cell covariate ondisc matrix. The matrix can be either an integer-valued expression matrix or a logical matrix of grna-to-cell assignments}

\item{grna_group}{the grna group (or vector of grna groups) to load}

\item{threshold}{the threshold to use for assigning grnas to cells (ignored if \code{covariate_odm} is a logical matrix)}

\item{grna_group_name}{the name of the column of the feature covariate matrix containing the grna group information}
}
\value{
a matrix of grouped grna-to-cell assignments
}
\description{
Loads data from an ondisc matrix into memory after thresholding and grouping grnas.
}
\examples{
# Please install the `ondiscdata` package before running the examples.
# install.packages("devtools")
# devtools::install_github("Katsevich-Lab/ondiscdata")

# EXAMPLE 1: an integer-valued matrix of grna expressions
odm_fp <- system.file("extdata", "odm/grna_expression/matrix.odm", package = "ondiscdata")
metadata_fp <- system.file("extdata", "odm/grna_expression/metadata.rds", package = "ondiscdata")
odm <- read_odm(odm_fp, metadata_fp)
indics <- load_thresholded_and_grouped_grna(odm, "grna_group_5")
indics <- load_thresholded_and_grouped_grna(odm, c("grna_group_1", "grna_group_11"))

# EXAMPLE 2: a logical matrix of grna indicators
odm_fp <- system.file("extdata", "odm/grna/matrix.odm", package = "ondiscdata")
metadata_fp <- system.file("extdata", "odm/grna/metadata.rds", package = "ondiscdata")
odm <- read_odm(odm_fp, metadata_fp)
# append grna group information -- three grnas/group
odm <- odm |>
mutate_feature_covariates(grna_group = rep(paste0("grna_group_", seq(1, nrow(odm)/3)), each = 3))
indics <- load_thresholded_and_grouped_grna(odm, c("grna_group_21", "grna_group_6"))
# alter grna group information -- one grna/group
odm <- odm |>
mutate_feature_covariates(grna_group = paste0("grna_group_", seq(1, nrow(odm))))
indics <- load_thresholded_and_grouped_grna(odm, "grna_group_2")
}
