% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crispr_screen.R
\name{load_thresholded_and_grouped_gRNA}
\alias{load_thresholded_and_grouped_gRNA}
\title{Load thresholded and grouped gRNA}
\usage{
load_thresholded_and_grouped_gRNA(
  covariate_odm,
  gRNA_group,
  threshold = 3,
  gRNA_group_name = "gRNA_group"
)
}
\arguments{
\item{covariate_odm}{a gRNA-by-cell covariate ondisc matrix. The matrix can be either an integer-valued expression matrix or a logical matrix of gRNA-to-cell assignments}

\item{gRNA_group}{the gRNA group (or vector of gRNA groups) to load}

\item{threshold}{the threshold to use for assigning gRNAs to cells (ignored if \code{covariate_odm} is a logical matrix)}

\item{gRNA_group_name}{the name of the column of the feature covariate matrix containing the gRNA group information}
}
\value{
a matrix of grouped gRNA-to-cell assignments
}
\description{
Loads data from an ondisc matrix into memory after thresholding and grouping gRNAs.
}
\examples{
# Please install the `ondiscdata` package before running the examples.
# install.packages("devtools")
# devtools::install_github("Katsevich-Lab/ondiscdata")

# EXAMPLE 1: an integer-valued matrix of gRNA expressions
odm_fp <- system.file("extdata", "odm/gRNA_expression/matrix.odm", package = "ondiscdata")
metadata_fp <- system.file("extdata", "odm/gRNA_expression/metadata.rds", package = "ondiscdata")
odm <- read_odm(odm_fp, metadata_fp)
indics <- load_thresholded_and_grouped_gRNA(odm, "gRNA_group_5")
indics <- load_thresholded_and_grouped_gRNA(odm, c("gRNA_group_1", "gRNA_group_11"))

# EXAMPLE 2: a logical matrix of gRNA indicators
odm_fp <- system.file("extdata", "odm/gRNA/matrix.odm", package = "ondiscdata")
metadata_fp <- system.file("extdata", "odm/gRNA/metadata.rds", package = "ondiscdata")
odm <- read_odm(odm_fp, metadata_fp)
# append gRNA group information -- three gRNAs/group
odm <- odm |>
mutate_feature_covariates(gRNA_group = rep(paste0("gRNA_group_", seq(1, nrow(odm)/3)), each = 3))
indics <- load_thresholded_and_grouped_gRNA(odm, c("gRNA_group_21", "gRNA_group_6"))
# alter gRNA group information -- one gRNA/group
odm <- odm |>
mutate_feature_covariates(gRNA_group = paste0("gRNA_group_", seq(1, nrow(odm))))
indics <- load_thresholded_and_grouped_gRNA(odm, "gRNA_group_2")
}
