% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s4_classes.R
\name{multimodal_ondisc_matrix}
\alias{multimodal_ondisc_matrix}
\title{\code{multimodal_ondisc_matrix} class}
\usage{
multimodal_ondisc_matrix(covariate_ondisc_matrix_list)

multimodal_ondisc_matrix(covariate_ondisc_matrix_list)
}
\arguments{
\item{covariate_ondisc_matrix_list}{a named list containing \code{covariate_ondisc_matrix} objects; the names are taken to be the names of the modalities.}
}
\value{
a multimodal_ondisc_matrix
}
\description{
A \code{multimodal_ondisc_matrix} represents multimodal data.

Construct a \code{multimodal_ondisc_matrix} from a list of \code{covariate_ondisc_matrix} objects.
}
\section{Slots}{

\describe{
\item{\code{modalities}}{a list containing \code{covariate_ondisc_matrix} objects representing different modalities.}

\item{\code{global_cell_covariates}}{a data frame containing the cell-specific covariates pooled across all modalities.}
}}

\examples{
# Please install the `ondiscdata` package before running the examples.
# install.packages("devtools")
# devtools::install_github("Katsevich-Lab/ondiscdata")

# Load odm from package
odm_gene_fp <- system.file("extdata", "odm/gene/matrix.odm", package = "ondiscdata")
metadata_gene_fp <- system.file("extdata", "odm/gene/metadata.rds", package = "ondiscdata")
odm_gene <- read_odm(odm_gene_fp, metadata_gene_fp)
odm_grna_fp <- system.file("extdata", "odm/grna/matrix.odm", package = "ondiscdata")
metadata_grna_fp <- system.file("extdata", "odm/grna/metadata.rds", package = "ondiscdata")
odm_grna <- read_odm(odm_grna_fp, metadata_grna_fp)

odm_multi <- multimodal_ondisc_matrix(list(gene = odm_gene, grna = odm_grna))
}
