% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s4_other_verbs.R
\name{mutate-covariates}
\alias{mutate-covariates}
\alias{mutate_cell_covariates}
\alias{mutate_feature_covariates}
\alias{mutate_cell_covariates,ondisc_matrix-method}
\alias{mutate_feature_covariates,ondisc_matrix-method}
\alias{mutate_cell_covariates,metadata_ondisc_matrix-method}
\alias{mutate_feature_covariates,metadata_ondisc_matrix-method}
\alias{mutate_cell_covariates,multimodal_ondisc_matrix-method}
\alias{mutate_feature_covariates,multimodal_ondisc_matrix-method}
\title{Mutate covariate matrices.}
\usage{
mutate_cell_covariates(x, ...)

mutate_feature_covariates(x, ...)

\S4method{mutate_cell_covariates}{ondisc_matrix}(x, ...)

\S4method{mutate_feature_covariates}{ondisc_matrix}(x, ...)

\S4method{mutate_cell_covariates}{metadata_ondisc_matrix}(x, ...)

\S4method{mutate_feature_covariates}{metadata_ondisc_matrix}(x, ...)

\S4method{mutate_cell_covariates}{multimodal_ondisc_matrix}(x, ...)

\S4method{mutate_feature_covariates}{multimodal_ondisc_matrix}(x, ...)
}
\arguments{
\item{x}{an object of class \code{ondisc_matrix}, \code{metadata_ondisc_matrix}, or \code{multimodal_ondisc_matrix}.}

\item{...}{arguments to dplyr::mutate}
}
\value{
an updated \code{ondisc} object
}
\description{
Functions to mutate cell-specific or feature-specific covariate matrix. Returns an \code{ondisc} object.
}
