% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/low_level_indexing.R
\name{return_spMatrix_from_index}
\alias{return_spMatrix_from_index}
\title{Return sparse matrix from index}
\usage{
return_spMatrix_from_index(
  h5_file,
  subset_vector,
  index_on_cell,
  logical_mat,
  underlying_dimension
)
}
\arguments{
\item{h5_file}{the file path to the h5 file}

\item{subset_vector}{an integer vector giving the subset}

\item{index_on_cell}{(boolean) index on cell (TRUE) or feature (FALSE)}

\item{logical_mat}{(boolean)}

\item{underlying_dimension}{a length 2 integer vector giving the underlying dimension of the expression matrix.}
}
\value{
a Matrix object.
}
\description{
Return sparse matrix from index
}
