% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/low_level_indexing.R
\name{return_spMatrix_from_index}
\alias{return_spMatrix_from_index}
\title{Return sparse matrix from index}
\usage{
return_spMatrix_from_index(h5_file, idx, col_idx)
}
\arguments{
\item{h5_file}{path to h5 file on disk}

\item{idx}{indexes to extract}

\item{col_idx}{boolean; extract column indexes (true) or row indexes (false)}
}
\value{
a sparse matrix of class Matrix.
}
\description{
Given an h5 file containing an on_disc_matrix, returns the sparse matrix corresponding to an index subset.
}
\examples{
h5_file <- system.file("extdata", "on_disc_matrix_1.h5", package = "ondisc")
if (h5_file != "") {
idx <- 2:3
col_idx <- TRUE
return_spMatrix_from_index(h5_file, idx, col_idx)
}
}
