% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s4_class.R
\name{dimnames,odm-method}
\alias{dimnames,odm-method}
\alias{rownames}
\title{Return the rownames of an \code{odm} object}
\usage{
\S4method{dimnames}{odm}(x)
}
\arguments{
\item{x}{an object of class \code{odm}}
}
\value{
the rownames of an \code{odm} object
}
\description{
\code{rownames()} returns the rownames (i.e., feature IDs) of an \code{odm} object.
}
\examples{
library(sceptredata)
directories_to_load <- paste0(
 system.file("extdata", package = "sceptredata"),
 "/highmoi_example/gem_group_", c(1, 2)
)
directory_to_write <- tempdir()
out_list <- create_odm_from_cellranger(
  directories_to_load = directories_to_load,
  directory_to_write = directory_to_write,
)
gene_odm <- out_list$gene
# return the rownames
rownames(gene_odm) |> head()
}
