% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io.R
\name{save_multimodal_odm}
\alias{save_multimodal_odm}
\title{Save a multimodal \code{ondisc_matrix}}
\usage{
save_multimodal_odm(multimodal_odm, multimodal_metadata_fp)
}
\arguments{
\item{multimodal_odm}{a \code{multimodal_ondisc_matrix} object}

\item{multimodal_metadata_fp}{path to multimodal metadata RDS file}
}
\description{
Save a multimodal \code{ondisc_matrix}
}
\examples{
# Please install the `ondiscdata` package before running the examples.
# install.packages("devtools")
# devtools::install_github("Katsevich-Lab/ondiscdata")

# Initialize a multimodal ODM
gene_odm_fp <- system.file("extdata", "odm/gene/matrix.odm", package = "ondiscdata")
odm_gene <- read_odm(gene_odm_fp, system.file("extdata", "odm/gene/metadata.rds", package = "ondiscdata"))
gRNA_odm_fp <- system.file("extdata", "odm/gRNA/matrix.odm", package = "ondiscdata")
odm_gRNA <- read_odm(gRNA_odm_fp, system.file("extdata", "odm/gRNA/metadata.rds", package = "ondiscdata"))
multimodal_odm <- multimodal_ondisc_matrix(list(gene = odm_gene, gRNA = odm_gRNA))
multimodal_metadata_fp <- paste0(create_new_directory(), "/multimodal_metadata.rds")
save_multimodal_odm(multimodal_odm, multimodal_metadata_fp)
}
