% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io.R
\name{save_odm}
\alias{save_odm}
\alias{read_odm}
\title{Save ODM}
\usage{
save_odm(odm, metadata_fp)

read_odm(odm_fp, name = NULL)
}
\arguments{
\item{odm}{a metadata_ondisc_matrix object or ondisc_matrix object}

\item{metadata_fp}{file path to metadata.rds object}

\item{odm_fp}{file path to an ondisc matrix (or metadata ondisc matrix)}
}
\description{
Saves an ondisc matrix or metadata ondisc matrix.
}
\examples{

}
