% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io.R
\name{save_odm}
\alias{save_odm}
\alias{read_odm}
\title{Save or read an \code{ondisc_matrix}}
\usage{
save_odm(odm, metadata_fp)

read_odm(odm_fp, metadata_fp = NULL)
}
\arguments{
\item{odm}{an ondisc_matrix object}

\item{metadata_fp}{path to metadata RDS file}

\item{odm_fp}{file path to a backing .odm file}
}
\value{
NULL for \code{save_odm}; a loaded \code{ondisc} object for \code{read_odm}
}
\description{
Functions to save and read \code{ondisc_matrix} objects.
}
\details{
The backing .odm file is a static, read-only file that is never altered or copied.
\code{read_odm} constructs an \code{ondisc_matrix} from a given backing .odm file and (optionally) a metadata file.
\code{save_odm} writes a new metadata file to disk (potentially overwriting a previous metadata file of the same name), leaving the backing .odm file unaltered.
}
\examples{
# Install the `ondiscdata` package to run the examples.
# devtools::install_github("timothy-barry/ondiscdata")

# Load odm from package
odm_fp <- system.file("extdata", "odm/gene/matrix.odm", package = "ondiscdata")
metadata_fp <- system.file("extdata", "odm/gene/metadata.rds", package = "ondiscdata")
odm <- read_odm(odm_fp, metadata_fp)
# assign a new ondisc_matrix by subsetting the original
odm_subset <- odm[1:10,]
# save the subsetted odm, and remove it from memory
tempfile <- create_new_directory()
metadata_subset_fp <- paste0(tempfile, "/metadata_subset.rds")
save_odm(odm_subset, metadata_subset_fp)
# finally, load the subsetted odm
odm_subset <- read_odm(odm_fp, metadata_subset_fp)
}
