% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s4_classes_methods.R
\name{show,ondisc_matrix-method}
\alias{show,ondisc_matrix-method}
\title{Print basic information to console}
\usage{
\S4method{show}{ondisc_matrix}(object)
}
\arguments{
\item{object}{an on_dist_matrix to show}
}
\description{
Also display object class, number of rows, and number of columns.
}
\examples{
# NOTE: You must create the HDF5 file "example.h5" to run this example.
# Navigate to the help file of "create_on_disc_matrix_from_10x_mtx"
# (via ?create_on_disc_matrix_from_10x_mtx), and execute the code in the example.
odm_fp <- system.file("extdata", "example.h5", package = "ondisc")
if (odm_fp != "") { # if required file exists, ...
odm <- ondisc_matrix(h5_file = odm_fp)
show(odm)
}
}
