% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s4_get_basic_info_methods.R
\name{show}
\alias{show}
\alias{show,ondisc_matrix-method}
\alias{show,metadata_ondisc_matrix-method}
\alias{show,multimodal_ondisc_matrix-method}
\title{Print basic information to the console}
\usage{
\S4method{show}{ondisc_matrix}(object)

\S4method{show}{metadata_ondisc_matrix}(object)

\S4method{show}{multimodal_ondisc_matrix}(object)
}
\arguments{
\item{object}{an object of class ondisc_matrix, covaraite_ondisc_matrix, or multimodal_ondisc_matrix}
}
\value{
NULL; called for printing
}
\description{
Print basic information to the console
}
