% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s4_class.R
\name{[,odm,ANY,missing,missing-method}
\alias{[,odm,ANY,missing,missing-method}
\title{Load a row of an \code{odm} into memory}
\usage{
\S4method{[}{odm,ANY,missing,missing}(x, i, j, drop)
}
\arguments{
\item{x}{an object of class \code{odm}}

\item{i}{the index of the row to load into memory. \code{i} can be either an integer (specifying the integer-based index of the row to load into memory) or a string (specifying the row to load into memory by feature ID).}

\item{j}{not used}

\item{drop}{not used}
}
\value{
a (numeric) expression vector
}
\description{
The operator \code{[} can be called on an \code{odm} to load a row of an \code{odm} into memory.
}
\examples{
library(sceptredata)
directories_to_load <- paste0(
 system.file("extdata", package = "sceptredata"),
 "/highmoi_example/gem_group_", c(1, 2)
)
directory_to_write <- tempdir()
out_list <- create_odm_from_cellranger(
  directories_to_load = directories_to_load,
  directory_to_write = directory_to_write,
)
gene_odm <- out_list$gene
# extract rows into memory by index and ID
v1 <- gene_odm[10L,]
v2 <- gene_odm["ENSG00000173825",]
}
