% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s4_classes_methods.R
\name{subset-covariate-odm}
\alias{subset-covariate-odm}
\alias{[,metadata_ondisc_matrix,ANY,ANY,missing-method}
\alias{[,metadata_ondisc_matrix,ANY,missing,missing-method}
\alias{[,metadata_ondisc_matrix,missing,ANY,missing-method}
\alias{[,metadata_ondisc_matrix,missing,missing,missing-method}
\title{Subset a \code{metadata_ondisc_matrix} with \code{[}}
\usage{
\S4method{[}{metadata_ondisc_matrix,ANY,ANY,missing}(x, i, j, drop)

\S4method{[}{metadata_ondisc_matrix,ANY,missing,missing}(x, i, j, drop)

\S4method{[}{metadata_ondisc_matrix,missing,ANY,missing}(x, i, j, drop)

\S4method{[}{metadata_ondisc_matrix,missing,missing,missing}(x, i, j, drop)
}
\arguments{
\item{x}{A metadata_ondisc_matrix object}

\item{i}{Vector (numeric, logical, or character) indicating features to keep}

\item{j}{Vector (numeric, logical, or character) indicating cells to keep}

\item{drop}{not used}
}
\value{
A subset metadata_ondisc_matrix object.
}
\description{
The user can pass logical, character, of numeric vectors to \code{`[`}. Character vectors correspond to feature IDs (for rows) and cell barcodes (for columns).
}
\examples{
# NOTE: You must create the HDF5 file "expressions.h5" and the RDS file
# "expressions.rds" to run this example. Navigate to the help file of
# "create_ondisc_matrix_from_mtx" (via ?create_ondisc_matrix_from_mtx),
# and execute the code in the first code block.
expressions_fp <- paste0(tempdir(), "/expressions.rds")
if (file.exists(expressions_fp)) {
expressions <- readRDS(expressions_fp)
# keep cells 100-110
x <- expressions[,100:110]
# keep the following genes
x <- expressions[c("ENSG00000188305", "ENSG00000257284", "ENSG00000251655"),]
}
}
