% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s4_subset_extract.R
\name{subset-odm}
\alias{subset-odm}
\alias{[,ondisc_matrix,missing,missing,missing-method}
\alias{[,ondisc_matrix,ANY,missing,missing-method}
\alias{[,ondisc_matrix,missing,ANY,missing-method}
\alias{[,ondisc_matrix,ANY,ANY,missing-method}
\alias{[,metadata_ondisc_matrix,ANY,ANY,missing-method}
\alias{[,metadata_ondisc_matrix,ANY,missing,missing-method}
\alias{[,metadata_ondisc_matrix,missing,ANY,missing-method}
\alias{[,metadata_ondisc_matrix,missing,missing,missing-method}
\alias{[,multimodal_ondisc_matrix,missing,missing,missing-method}
\alias{[,multimodal_ondisc_matrix,missing,ANY,missing-method}
\alias{[,multimodal_ondisc_matrix,ANY,ANY,ANY-method}
\title{Subset using the \code{[} operator.}
\usage{
\S4method{[}{ondisc_matrix,missing,missing,missing}(x, i, j, drop)

\S4method{[}{ondisc_matrix,ANY,missing,missing}(x, i, j)

\S4method{[}{ondisc_matrix,missing,ANY,missing}(x, i, j)

\S4method{[}{ondisc_matrix,ANY,ANY,missing}(x, i, j)

\S4method{[}{metadata_ondisc_matrix,ANY,ANY,missing}(x, i, j, drop)

\S4method{[}{metadata_ondisc_matrix,ANY,missing,missing}(x, i, j, drop)

\S4method{[}{metadata_ondisc_matrix,missing,ANY,missing}(x, i, j, drop)

\S4method{[}{metadata_ondisc_matrix,missing,missing,missing}(x, i, j, drop)

\S4method{[}{multimodal_ondisc_matrix,missing,missing,missing}(x, i, j, drop)

\S4method{[}{multimodal_ondisc_matrix,missing,ANY,missing}(x, i, j, drop)

\S4method{[}{multimodal_ondisc_matrix,ANY,ANY,ANY}(x, i, j, drop)
}
\arguments{
\item{x}{an \code{ondisc_matrix}, \code{metadata_ondisc_matrix}, or \code{multimodal_ondisc_matrix} object.}

\item{i}{a vector (numeric, logical, or character) indicating features to keep.}

\item{j}{a vector (numeric, logical, or character) indicating cells to keep.}

\item{drop}{not used}
}
\value{
An appropriately subset object of the same class as \code{x}.
}
\description{
Apply the \code{[} operator to an \code{ondisc_matrix}, \code{metadata_ondisc_matrix}, or \code{multimodal_ondisc_matrix}
to subset the object. You can pass logical, character, or numeric vectors to \code{[}; character
vectors are assumed to refer to feature IDs (for rows) and cell barcodes (for columns).
}
\details{
You can subset an \code{ondisc_matrix} and a \code{metadata_ondisc_matrix} by cell and/or feature. You can subset a
\code{multimodal_ondisc_matrix} by cell only (because the features differ across modalities).
}
