% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s4_subset_extract.R
\name{subset-odm}
\alias{subset-odm}
\alias{[,ondisc_matrix,missing,missing,missing-method}
\alias{[,ondisc_matrix,ANY,missing,missing-method}
\alias{[,ondisc_matrix,missing,ANY,missing-method}
\alias{[,ondisc_matrix,ANY,ANY,missing-method}
\alias{[,covariate_ondisc_matrix,ANY,ANY,missing-method}
\alias{[,covariate_ondisc_matrix,ANY,missing,missing-method}
\alias{[,covariate_ondisc_matrix,missing,ANY,missing-method}
\alias{[,covariate_ondisc_matrix,missing,missing,missing-method}
\alias{[,multimodal_ondisc_matrix,missing,missing,missing-method}
\alias{[,multimodal_ondisc_matrix,missing,ANY,missing-method}
\alias{[,multimodal_ondisc_matrix,ANY,ANY,ANY-method}
\title{Subset using the \code{[} operator.}
\usage{
\S4method{[}{ondisc_matrix,missing,missing,missing}(x, i, j, drop)

\S4method{[}{ondisc_matrix,ANY,missing,missing}(x, i, j)

\S4method{[}{ondisc_matrix,missing,ANY,missing}(x, i, j)

\S4method{[}{ondisc_matrix,ANY,ANY,missing}(x, i, j)

\S4method{[}{covariate_ondisc_matrix,ANY,ANY,missing}(x, i, j, drop)

\S4method{[}{covariate_ondisc_matrix,ANY,missing,missing}(x, i, j, drop)

\S4method{[}{covariate_ondisc_matrix,missing,ANY,missing}(x, i, j, drop)

\S4method{[}{covariate_ondisc_matrix,missing,missing,missing}(x, i, j, drop)

\S4method{[}{multimodal_ondisc_matrix,missing,missing,missing}(x, i, j, drop)

\S4method{[}{multimodal_ondisc_matrix,missing,ANY,missing}(x, i, j, drop)

\S4method{[}{multimodal_ondisc_matrix,ANY,ANY,ANY}(x, i, j, drop)
}
\arguments{
\item{x}{an \code{ondisc_matrix}, \code{covariate_ondisc_matrix}, or \code{multimodal_ondisc_matrix} object.}

\item{i}{a vector (numeric, logical, or character) indicating features to keep.}

\item{j}{a vector (numeric, logical, or character) indicating cells to keep.}

\item{drop}{not used}
}
\value{
An appropriately subset object of the same class as \code{x}.
}
\description{
Apply the \code{[} operator to an \code{ondisc_matrix}, \code{covariate_ondisc_matrix}, or \code{multimodal_ondisc_matrix}
to subset the object. You can pass logical, character, or numeric vectors to \code{[}; character
vectors are assumed to refer to feature IDs (for rows) and cell barcodes (for columns).
}
\details{
You can subset an \code{ondisc_matrix} and a \code{covariate_ondisc_matrix} by cell and/or feature. You can subset a
\code{multimodal_ondisc_matrix} by cell only (because the features differ across modalities).
}
\examples{
# Install the `ondiscdata` package to run the examples.
# devtools::install_github("timothy-barry/ondiscdata")

####################################
# EXAMPLE 1: covariate_ondisc_matrix
####################################
odm_fp <- system.file("extdata", "odm/gene/matrix.odm", package = "ondiscdata")
metadata_fp <- system.file("extdata", "odm/gene/metadata.rds", package = "ondiscdata")
odm <- read_odm(odm_fp, metadata_fp)

# numeric vector
odm_numeric <- odm[1:10, 1:10]
# logical vector
set.seed(1)
nrow_logical <- sample(x = c(TRUE, FALSE), size = nrow(odm), replace = TRUE)
ncol_logical <- sample(x = c(TRUE, FALSE), size = ncol(odm), replace = TRUE)
odm_logical <- odm[nrow_logical, ncol_logical]
# character vectors for feature IDs and cell barcodes
feature_ids_subset <- get_feature_ids(odm)[1:10]
cell_barcodes_subset <- get_cell_barcodes(odm)[1:10]
odm_char <- odm[feature_ids_subset,cell_barcodes_subset]

#####################################
# EXAMPLE 2: multimodal_ondisc_matrix
#####################################
odm_gene_fp <- system.file("extdata", "odm/gene/matrix.odm", package = "ondiscdata")
metadata_gene_fp <- system.file("extdata", "odm/gene/metadata.rds", package = "ondiscdata")
odm_gene <- read_odm(odm_gene_fp, metadata_gene_fp)
odm_grna_fp <- system.file("extdata",
"odm/grna_assignment/matrix.odm", package = "ondiscdata")
metadata_grna_fp <- system.file("extdata",
"odm/grna_assignment/metadata.rds", package = "ondiscdata")
odm_grna <- read_odm(odm_grna_fp, metadata_grna_fp)
odm_multi <- multimodal_ondisc_matrix(list(gene = odm_gene, grna = odm_grna))

# numeric vector
odm_multi_numeric <- odm_multi[, 1:10]
# logical vector
set.seed(1)
ncol_logical <- sample(x = c(TRUE, FALSE), size = ncol(odm_multi), replace = TRUE)
odm_multi_logical <- odm_multi[, ncol_logical]
# character vectors for cell barcodes
cell_barcodes_subset <- get_cell_barcodes(odm_multi)[1:10]
odm_multi_cell_barcodes <- odm[,cell_barcodes_subset]
}
