% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_creation_functs.R
\name{write_example_cellranger_dataset}
\alias{write_example_cellranger_dataset}
\title{Write example cellranger dataset}
\usage{
write_example_cellranger_dataset(
  n_features,
  n_cells,
  n_batch,
  modalities,
  directory_to_write,
  p_zero = NULL,
  p_set_col_zero = NULL,
  p_set_row_zero = NULL
)
}
\arguments{
\item{n_features}{number of features per matrix}

\item{n_cells}{number of cells}

\item{n_batch}{number of batches}

\item{modalities}{a character vector indicating the modalities to create. The vector should contain the elements "gene", "grna", or "protein".}

\item{p_zero}{(optional) the fraction of entries to set randomly to zero}

\item{p_set_col_zero}{(optional) the fraction of columns to set randomly to zero}

\item{p_set_row_zero}{(optional) the fraction of rows to set randomly to zero}

\item{dir_to_write}{the directory in which to write the files}
}
\value{
a list containing (i) a list of the synthetic expression matrices, (ii) a character vector containing the names of the genes (or NULL if a gene modality was not created), and (iii) a factor vector specifying the batch of each cell
}
\description{
\code{write_example_cellranger_dataset()} creates an example Cell Ranger dataset.
}
\examples{
n_features <- c(1000, 40, 400)
modalities <- c("gene", "protein", "grna")
n_cells <- 10000
n_batch <- 2
dir_to_write <- tempdir()
p_set_col_zero <- 0
out <- write_example_cellranger_dataset(
  n_features = n_features,
  n_cells = n_cells,
  n_batch = n_batch,
  modalities = modalities,
  dir_to_write = dir_to_write,
  p_set_col_zero = p_set_col_zero
)
}
